% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{ae_val}
\alias{ae_val}
\title{Title}
\usage{
ae_val(ae_table, treatment, value)
}
\arguments{
\item{ae_table}{a table with columns \code{treatment},
\code{ae} (the names of different adverse events), \code{prob}
(the probability of the given adverse event occurring under the
given treatment),
and others naming values that can be represented, for example
cost or disutility.}

\item{treatment}{the treatment for which the value is desired}

\item{value}{the specific value desired}
}
\value{
the requested value for the requested treatment
}
\description{
Title
}
\details{
The underlying function is memoised for efficiency.
}
\examples{
AEs <- data.frame(treatment = c("A", "A", "B", "B"),
                  ae = c("ae1", "ae2", "ae1", "ae2"),
                  prob = c(0.1, 0.1, 0.2, 0),
                  cost = c(100, 200, 100, 200))
ae_val(AEs, "A", "cost")
ae_val(AEs, "B", "cost")

}
