% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{age_to_time}
\alias{age_to_time}
\alias{age_to_time.surv_table}
\alias{age_to_time.surv_object}
\alias{age_to_time.surv_pooled}
\alias{age_to_time.default}
\title{Convert an age-based table to a cycle-based table}
\usage{
age_to_time(x, ...)

\method{age_to_time}{surv_table}(x, age_init, cycle_length)

\method{age_to_time}{surv_object}(x, ...)

\method{age_to_time}{surv_pooled}(x, ...)

\method{age_to_time}{default}(x, ...)
}
\arguments{
\item{x}{a survival object}

\item{age_init}{initial age the table will be applied to}

\item{cycle_length}{length of cycles}
}
\value{
a table with \code{time} renormalized to be in terms of cycles,
with time 0 corresponding to age \code{age_init}
}
\description{
Convert an age-based table to a cycle-based table
}
\details{
This is used to convert a mortality table for use as a
survival curve in terms of cycles of the model.  Typically,
age_surv will have been constructed using \code{\link[=define_surv_table]{define_surv_table()}}
or \code{\link[=define_wtd_surv_table]{define_wtd_surv_table()}}.
}
