% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{apply_shift}
\alias{apply_shift}
\title{Apply a time shift}
\usage{
apply_shift(dist, shift)
}
\arguments{
\item{dist}{A survival distribution.}

\item{shift}{A time shift to be applied.}
}
\value{
A \code{surv_shift} object.
}
\description{
Shift a survival distribution in time.
}
\examples{

dist1 <- define_survival(distribution = "exp", rate = .25)
shift_dist <- apply_shift(dist1, 4)
}
