% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{as_expr_list}
\alias{as_expr_list}
\title{Convert Lazy Dots to Expression List}
\usage{
as_expr_list(.dots)
}
\arguments{
\item{.dots}{A lazy dots object.}
}
\value{
A list of expression.
}
\description{
This function is used by \code{\link{interp_heemod}}.
}
\keyword{internal}

