% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{calc_surv_from_prob}
\alias{calc_surv_from_prob}
\title{Calculate Probability of Survival}
\usage{
calc_surv_from_prob(x)
}
\arguments{
\item{x}{A vector of per-cycle event probabilities.}
}
\value{
The survival probabilities.
}
\description{
Calculates the probability of survival from a vector of event
probabilities
}
\keyword{internal}
