% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model_index}
\alias{check_model_index}
\title{Check Model Index}
\usage{
check_model_index(x, i)
}
\arguments{
\item{x}{A result from \code{\link{run_models}}.}

\item{i}{A model index, character or numeric.}
}
\value{
Nothing, just throws an error if an incorrect model 
index is used.
}
\description{
Check Model Index
}

