% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_names}
\alias{check_names}
\title{Check Names}
\usage{
check_names(x)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
Nothing, just throws an error if a reserved name
  is encountered.
}
\description{
Throws an error if any of the names are reserved.
}
\details{
Reserved names are \code{markov_cycle} and anything
starting with \code{.}.
}
\keyword{internal}

