% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{check_states}
\alias{check_states}
\title{Check Model States for Consistency}
\usage{
check_states(x)
}
\arguments{
\item{x}{An object of class \code{uneval_states}.}
}
\value{
\code{NULL}
}
\description{
For internal use.
}
\details{
All states should have the same value names.
}
\keyword{internal}

