% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_strategy_index}
\alias{check_strategy_index}
\title{Check Strategy Index}
\usage{
check_strategy_index(x, i, allow_multiple = FALSE)
}
\arguments{
\item{x}{A result from \code{\link{run_model}}.}

\item{i}{A strategy index, character or numeric.}

\item{allow_multiple}{logical. Allow multiple strategy
index?}
}
\value{
Strategy names.
}
\description{
Check Strategy Index
}
\keyword{internal}

