% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_factors}
\alias{clean_factors}
\title{Convert Data Frame Factor Variables to Character}
\usage{
clean_factors(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame.
}
\description{
Convert Data Frame Factor Variables to Character
}
\keyword{internal}

