% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_for_heemod.R
\name{cluster_for_heemod}
\alias{cluster_for_heemod}
\title{Prepare a cluster to use with heemod}
\usage{
cluster_for_heemod(num_cores)
}
\arguments{
\item{num_cores}{The number of cores desired}
}
\value{
If num_cores is 1 or 0, NULL.   If num_cores is a
  larger integer, a cluster loaded with the appropriate
  packages.
}
\description{
Prepare a cluster to use with heemod
}

