% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_models}
\alias{combine_models}
\title{Run Models on Demographic Data}
\usage{
combine_models(list_newmodels, weights, oldmodel)
}
\arguments{
\item{x}{Result from \code{\link{run_models}}.}

\item{demographics}{A table containing demographic
data (see details).

The demographic data table must be a \code{data.frame} with
the following properties: the column names must be parameter 
names used in \code{\link{define_parameters}}; and an optional
column \code{.weights} can give the respective weight of
each row in the target population.

Weights are automatillcally scaled. If no weights are provided
equal weights are used for each strata.}
}
\value{
A \code{run_demographics} object, mostly
similar to a result from \code{\link{run_models}}. \code{plot} and
\code{summary} methods are available.
}
\description{
Run a model on a set of demographic data giving population
characteristics per strata, return aggregated results
to estimate population-level values.
}

