% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{combine_part_surv}
\alias{combine_part_surv}
\alias{combine_part_surv_}
\title{recombine elements from different part_surv ojects}
\usage{
combine_part_surv(fit_matrix, ...)

combine_part_surv_(fit_matrix, choices)
}
\arguments{
\item{fit_matrix}{a matrix of partitioned survival
(\code{part_surv}) objects, generally
from \code{partitioned_survival_from_tabular}.
Each row corresponding to a distribution and each column
to a group or condition for which a partitioned survival object
was created.}

\item{...}{additional named arguments when using \code{combine_part_surv};
see below.}

\item{choices}{a list specifying which fits to use; see below.}
}
\value{
a list of partitioned survival objects
}
\description{
recombine elements from different part_surv ojects
}
\details{
The \code{choices} argument in \code{choose_part_surv_} and the
\code{...} argument in \code{choose_part_surv} specify which elements of which fit
survival objects to combine.
\code{choices} is a list, where the name of each element should be the name of
one of the strategies.   Each of those elements should be a list with elements
named pfs and os, for
the two elements of a partial survival object, and each element
must be a distribution type (a distribution used in the survival
fit).   Each name should be one of the groups
or subgroups; that is, a column name of \code{fit_matrix}.
}
\examples{
fit_matrix <- 
  partitioned_survival_from_tabular(base_dir = system.file("tabular\\\\surv", 
                                                package = "heemod"), 
                                  ref_file = "example_oncSpecs.csv", 
                                  df_env = new.env(), 
                                  state_names = c("ProgressionFree", "Progressive", 
                                  "Terminal", "Death"), 
                                  save_fits = FALSE,
                                  just_load = FALSE)
## create new partitioned survival objects.   For group A,
## use the exponential fit for progression-free survival, and
## the Weibull fit for overall survival; for B, use the exponential
## and lognormal fits respectively.
combine_part_surv(fit_matrix, A = list(pfs = "exp", os = "weibull"), 
                              B = list(pfs = "exp", os = "lnorm"))
combine_part_surv_(fit_matrix, 
                  choices = list(A = list(pfs = "exp", os = "weibull"), 
                                 B = list(pfs = "exp", os = "lnorm")))
}
