% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{combine_part_surv_from_fit_file}
\alias{combine_part_surv_from_fit_file}
\title{Create partial survival objects from fit files}
\usage{
combine_part_surv_from_fit_file(file_name, obj_name, ...)
}
\arguments{
\item{file_name}{loadable .RData file containing a fit matrix.}

\item{obj_name}{the name of the fit matrix inside the .RData file.}

\item{...}{additional arguments for \code{\link[=combine_part_surv]{combine_part_surv()}}.}
}
\value{
a list of partitioned survival objects
}
\description{
Create partial survival objects from fit files
}
