% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{combine_probs}
\alias{combine_probs}
\title{Combine Probabilities}
\usage{
combine_probs(...)
}
\arguments{
\item{...}{Probability vectors.}
}
\value{
A probability vector.
}
\description{
Given several independent probabilities of an event,
return the final probability of the event.
}
\details{
This function is only correct if the probabilities are
independent!
}
\examples{

(p1 <- runif(5))
(p2 <- runif(5))
combine_probs(p1, p2)

}
