% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Title}
\usage{
compute_covariance(x, model = 1, linear = TRUE, scale_x = TRUE,
  scale_y = FALSE)
}
\arguments{
\item{x}{Result from \code{\link{run_probabilistic}}.}

\item{model}{Name or index of model.}

\item{linear}{Look at linear relations?}

\item{scale_x}{Scale inputs?}

\item{scale_y}{Scale outputs?}
}
\value{
A data frame.
}
\description{
Title
}

