% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_summary.R
\name{compute_icer}
\alias{compute_icer}
\title{Compute ICER}
\usage{
compute_icer(x, strategy_order = order(x$.effect), threshold = 30000)
}
\arguments{
\item{x}{Result of \code{\link[=run_model]{run_model()}}.}

\item{strategy_order}{Order in which the strategies
should be sorted. Default: by increasing effect.}

\item{threshold}{ICER threshold for net monetary benefit
computation.}
}
\value{
A \code{data.frame} with computed ICER.
}
\description{
Compute ICER for Markov models.
}
\details{
Models are ordered by effectiveness and ICER are computed
sequentially.
}
\keyword{internal}
