% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{compute_surv}
\alias{compute_surv}
\title{Evaluate Survival Distributions}
\usage{
compute_surv(x, time, cycle_length = 1, type = c("prob", "survival"), ...)
}
\arguments{
\item{x}{A survival distribution object}

\item{time}{The \code{model_time} or \code{state_time} for which
to predict.}

\item{cycle_length}{The value of a Markov cycle in
absolute time units.}

\item{type}{either \code{prob}, for transition probabilities,
or \code{surv}, for survival}

\item{...}{arguments passed to methods.}
}
\value{
Returns either the survival probalities or
conditional probabilities of event for each cycle.
}
\description{
Generate either survival probabilities or conditional
probabilities of event for each model cycle.
}
\details{
The results of \code{compute_surv()} are memoised for
\code{options("heemod.memotime")} (default: 1 hour) to
increase resampling performance.
}
