% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_survival.R
\name{construct_survival}
\alias{construct_survival}
\title{construct a survival object from tabular specification}
\usage{
construct_survival(surv_def, fit_matrix_name = "fit_matrix", env = NULL)
}
\arguments{
\item{surv_def}{a data frame with the specification.  See details.}

\item{fit_matrix_name}{the name of the matrix from which to take fits.
Typically produced by \code{\link[=survival_fits_from_tabular]{survival_fits_from_tabular()}}.}
}
\value{
a list with one element for each strategy.   Each element
is in turn a list with two elements, pfs and os.   And those
elements are survival objects of various kinds, with the
commonality that they can be used in \code{\link[=compute_surv]{compute_surv()}}.
}
\description{
construct a survival object from tabular specification
}
\details{
columns:  .strategy, .type, dist, until
dist can be either the name of a distribution
along with parameters, or a reference to a fit
for example:  fit('exp') or exp(rate = 0.5)
}
