% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_timescale_of_parameters.R
\name{convert_timescale_of_parameters}
\alias{convert_timescale_of_parameters}
\title{Title  Re-scales the input in time, converting the value between two time-spans.}
\usage{
convert_timescale_of_parameters(input, scale_from, scale_to)
}
\arguments{
\item{input}{name or value of a parameter that is to be re-scaled  from the original time scale \code{scale_from} to the new scale \code{scale_to}.}

\item{scale_from}{length of the original time period}

\item{scale_to}{length of the original time period}
}
\value{
value scaled under the assumption of *compound* discounting (read here: https://en.wikipedia.org/wiki/Compound_interest).
}
\description{
Title  Re-scales the input in time, converting the value between two time-spans.
}
\details{
Some input parameters may have different time scale than is the model time-step, 
  e.g. discount rates for costs and benefits are typically annual, while the model time-step 
  is likely to be shorter. To assure the discounting is correct, the correctly scaled input values
  must be provided to the calculation.
  Users may use this function to convert the inputs into the horizon of the model time-step.
  Continuous discounting is assumed, i.e. when converting a long-term discount rate into a short-term rate, 
  we assume that a partial gain from one short term is multiplicatively discounted in all following
  short terms. This means, e.g. that for a monthly rate of 1% we get the annual rate of 12.6825% (i.e. 
  (1+0.01)^12 - 1; example from https://en.wikipedia.org/wiki/Compound_interest).
  At the same time, we assume the short-term rate is time-invariant.
  Ad time units: we use only ratio of the two scales; thus, the time unit is arbitrary.
}
\examples{
  ## 1\% monthly interest rate to annual
  convert_timescale_of_parameters(0.01, 1, 12)
  ## 3\% annual discount rate to (approximately) weekly 
  convert_timescale_of_parameters(0.03, 52, 1)
}

