% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_counts.R
\name{correct_counts}
\alias{correct_counts}
\title{Title}
\usage{
correct_counts(x, method = c("life-table", "beginning", "end"))
}
\arguments{
\item{x}{a count table}

\item{method}{either a string or a function - see details.}
}
\value{
a data frame with one fewer row than \code{x}.
}
\description{
Title
}
\details{
If \code{method} is a string, it should be either \code{life-table},
\code{beginning}, or \code{end}.  if \code{method = 'life-table'}, then the function
returns, as the count for
a cycle, the average of the count at the beginning and at the end
of the cycle.   If \code{method = 'beginning'``, it considers transitions to take place at the beginning of each cycle, so it returns, for each cycle, the number of counts at the end of the cycle (or, equivalently, at the beginning of the next cycle). If}method = 'end'`, it considers transitions to take place at the end
of each cycle, so it returns, for each cycle, the number of counts at
the beginning of the cycle (or, equivalently, at the end of the previous
cycle).

If \code{method} is a function, the \code{x} will be passed to that function.
The function must return a data frame with one fewer row than \code{x}.
}
