% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{cost_iv_administration}
\alias{cost_iv_administration}
\title{Cost of administration for an intravenous treatment}
\usage{
cost_iv_administration(iv_time, cost_first_unit, cost_addl_units)
}
\arguments{
\item{iv_time}{The infusion time, in units compatible with the costs}

\item{cost_first_unit}{cost for the first time unit}

\item{cost_addl_units}{cost for the second time unit}
}
\value{
the cost
}
\description{
Cost of administration for an intravenous treatment
}
\examples{
cost_iv_administration(0.5, 100, 20) # = 50
cost_iv_administration(1.5, 100, 20) # = 110
}
