% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{cost_iv_compound_administration}
\alias{cost_iv_compound_administration}
\title{Cost of administration for an intravenous treatment}
\usage{
cost_iv_compound_administration(data_table, compound, time_col = "iv_time_hr",
  first_cost_col = "cost_admin_first_hr",
  add_cost_col = "cost_admin_addl_hr")
}
\arguments{
\item{data_table}{a data frame; see details}

\item{compound}{the name of the compound}
}
\value{
the cost
}
\description{
Cost of administration for an intravenous treatment
}
\details{
\code{data_table} must have columns \code{compound}, \code{param},
and \code{value}.   The required values are found in \code{data_table}
using \code{\link[=look_up]{look_up()}}.
}
\examples{
exampleParams <- 
  data.frame(compound = "X", 
  param = c("cost_admin_first_hr", "cost_admin_addl_hr",
            "iv_time_hr"),
  value = c(100, 20, 1.5))
  cost_iv_compound_administration(exampleParams, "X")
}
