% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_demographic_table}
\alias{create_demographic_table}
\title{Read a Demographic Table}
\usage{
create_demographic_table(newdata, params)
}
\arguments{
\item{newdata}{A data frame.}

\item{params}{Parameters of a model, to check that all
the columns in the demographic table (other than the
weight column) are in the model.}

\item{...}{catches other, unwanted arguments.}
}
\value{
A data frame.
}
\description{
This function mostly checks whether the parameters are
correct.
}
\details{
An optional \code{.weights} column can exist in the file.
}
\keyword{internal}

