% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_df_from_tabular}
\alias{create_df_from_tabular}
\title{Load Data From a Folder Into an Environment}
\usage{
create_df_from_tabular(df_dir, df_envir)
}
\arguments{
\item{df_dir}{A directory containing the files.}

\item{df_envir}{An environment.}
}
\value{
The environment with the data frames.
}
\description{
Reads files containing data frames (in tabular format) 
from a directory, and loads them in an environment to be 
available during an analysis.
}
\details{
The files must be in .csv, .xls, or .xlsx format. A file 
my_df.csv (or my_df.xls, or my_df.xlsx) will be loaded as
a data frame my_df.
}
\keyword{internal}

