% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tabular.R
\name{create_import_template}
\alias{create_import_template}
\title{Create a Tabular Data Template}
\usage{
create_import_template(path, model_names, state_names, state_value_names)
}
\arguments{
\item{path}{Path to an empty folder where
the template will be created.}

\item{model_names}{Names of models.}

\item{state_names}{Names of states.}

\item{state_value_names}{Names of state values.}
}
\value{
\code{NULL}
}
\description{
Create a Tabular Data Template
}

