% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_model_from_tabular}
\alias{create_model_from_tabular}
\title{Create a \code{heemod} Model From Tabular Files Info}
\usage{
create_model_from_tabular(state_info, tm_info, surv_info, strategy_name,
  df_env = globalenv())
}
\arguments{
\item{df_env}{An environment containing external data.}

\item{state_file}{A state tabular file (file path or 
parsed file).}

\item{tm_file}{A transition matrix tabular file (file 
path or parsed file).}
}
\value{
A \code{heemod} model as returned by 
  \code{\link{define_strategy}}.
}
\description{
Calls \code{\link{create_states_from_tabular}} and
\code{\link{create_matrix_from_tabular}}.
}
\keyword{internal}

