% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_model_list_from_tabular}
\alias{create_model_list_from_tabular}
\title{Read Models Specified by Files}
\usage{
create_model_list_from_tabular(ref, base_model, df_env)
}
\arguments{
\item{ref}{Imported reference file.}

\item{base_model}{Which of the models should be the base 
model.}
}
\value{
A list of unevaluated models.
}
\description{
Read Models Specified by Files
}

