% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_options_from_tabular}
\alias{create_options_from_tabular}
\title{Create Model Options From a Tabular Input}
\usage{
create_options_from_tabular(opt)
}
\arguments{
\item{opt}{An option data frame.}
}
\value{
A list of model options.
}
\description{
Create Model Options From a Tabular Input
}
\keyword{internal}

