% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_parameters_from_tabular}
\alias{create_parameters_from_tabular}
\title{Create a Parameter Definition From Tabular Input}
\usage{
create_parameters_from_tabular(param_defs, df_env = globalenv())
}
\arguments{
\item{param_defs}{A parameter definition file.}

\item{df_env}{An environment containing external data.}
}
\value{
The parameter definition.
}
\description{
If specified in the tabular file, DSA and PSA can also be
created.
}
\details{
The tabular parameter definition file can have the 
following columns: \code{parameter} (the parameter name, 
required), \code{value} (required), \code{low} and 
\code{high} (if both are present, a deterministic 
sensitivity analysis will be performed), and \code{psa} 
(a definition of a distribution to use in a probabilistic
sensitivity analysis. Other columns will be ignored.
}
\keyword{internal}

