% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{create_part_surv_from_tabular}
\alias{create_part_surv_from_tabular}
\title{Create a partitioned survival model object from tabular
input}
\usage{
create_part_surv_from_tabular(surv_dir, df_env, model_names)
}
\arguments{
\item{df_env}{an environment in which survival results
will be stored.}

\item{model_names}{names of the models.}

\item{ref}{Name of the reference file.}
}
\value{
a list, with one element for each model.  Each
  element of the list is also a list, with one element
  for PFS and one for OS.
}
\description{
Create a partitioned survival model object from tabular
input
}

