% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_define.R
\name{define_dsa}
\alias{define_dsa}
\alias{define_dsa_}
\title{Define a Sensitivity Analysis}
\usage{
define_dsa(...)

define_dsa_(par_names, low_dots, high_dots)
}
\arguments{
\item{...}{A list of parameter names and min/max values 
of the form \code{var1, min(var1), max(var1), var2,
min(var2), max(var2), ...}.}

\item{par_names}{String vector of parameter names.}

\item{low_dots, high_dots}{Used to work around
non-standard evaluation.}
}
\value{
A \code{sensitivity} object.
}
\description{
Define parameter variations for a Markov model 
sensitivity analysis.
}
\examples{

define_dsa(
  a, 10, 45,
  b, .5, 1.5
)

}

