% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_inflow}
\alias{define_inflow}
\title{Define Inflow for a BIA}
\usage{
define_inflow(...)
}
\arguments{
\item{...}{Name-value pairs of expressions definig
inflow.}
}
\value{
An object similar to the return value of
  \code{\link{define_parameters}}.
}
\description{
This function is a placeholder.
}
\details{
This function only takes constant values. Eventually
time-dependant expression will be accepted (with
model-time dependency only).
}
