% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_init}
\alias{define_init}
\alias{define_init_}
\title{Define Initial Counts}
\usage{
define_init(...)

define_init_(.dots)
}
\arguments{
\item{...}{Name-value pairs of expressions defining
initial counts.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object similar to the return value of
\code{\link[=define_parameters]{define_parameters()}}.
}
\description{
Define Initial Counts
}
