% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_initial_values}
\alias{define_initial_values}
\title{Define Initial Values by State and Strategy}
\usage{
define_initial_values(...)
}
\arguments{
\item{...}{Name-value pairs of expressions defining
initial values.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object similar to the return value of
\code{\link[=define_parameters]{define_parameters()}}.
}
\description{
Define Initial Values by State and Strategy
}
