% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{define_multinom}
\alias{define_multinom}
\title{Define That Parameters Belong to the Same Multinomial
Distribution}
\usage{
define_multinom(...)
}
\arguments{
\item{...}{A list of parameter names.}
}
\value{
An object of class \code{multinomial}.
}
\description{
Define That Parameters Belong to the Same Multinomial
Distribution
}
\examples{

define_multinom(a, b, c)

}

