% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{define_part_surv}
\alias{define_part_surv}
\title{Define Partitioned Survival}
\usage{
define_part_surv(pfs, os, state_names, km_limit)
}
\arguments{
\item{pfs}{Distribution of progression-free survival.}

\item{os}{Distribution of overall survival.}

\item{state_names}{character vector, optional, length 3. State
names for original state, progression and death
respectively.}

\item{km_limit}{}
}
\description{
Define a partitioned survival model with progression-free
survival and overall survival.
}

