% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{define_resample}
\alias{define_resample}
\title{Define parameters distribution for probabilistc analysis}
\usage{
define_resample(..., mat_cor = diag(length(list(...))))
}
\arguments{
\item{...}{Name-value pairs of expressions definig 
parameter distributions.}

\item{mat_cor}{A correlation matrix for parameters.}
}
\value{
An object of class \code{resamp_definition}.
}
\description{
Define the properties of parameter distributions and 
their correlation structure for probabilistic incertitude
analysis of Markov models.
}

