% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_define.R
\name{define_sensitivity}
\alias{define_sensitivity}
\alias{define_sensitivity_}
\title{Define a Sensitivity Analysis}
\usage{
define_sensitivity(...)

define_sensitivity_(.dots)
}
\arguments{
\item{...}{A named list of min and max values that 
parameters will take.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
A \code{sensitivity} object.
}
\description{
Define parameter variations for a Markov model 
sensitivity analysis.
}
\details{
Parameter variations windows are given as length 2 vector
of the form \code{c(min, max)}.
}
\examples{

define_sensitivity(
  a = c(10, 45),
  b = c(.5, 1.5)
)

}

