% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_starting_values}
\alias{define_starting_values}
\alias{define_starting_values_}
\title{Define Starting State Values}
\usage{
define_starting_values(...)

define_starting_values_(.dots)
}
\arguments{
\item{...}{Name-value pairs of expressions defining
starting values.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object similar to the return value of
\code{\link[=define_parameters]{define_parameters()}}.
}
\description{
Define Starting State Values
}
