% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{define_state}
\alias{define_state}
\alias{define_state_}
\alias{modify.state}
\title{Define a Markov Model State}
\usage{
define_state(..., count_types)

define_state_(.dots, count_types)

\method{modify}{state}(.OBJECT, ...)
}
\arguments{
\item{...}{Name-value pairs of expressions defining state
values.}

\item{count_types}{the count type with which each
value should be calculated.}

\item{.dots}{Used to work around non-standard evaluation.}

\item{.OBJECT}{An object of class \code{state}.}
}
\value{
An object of class \code{state} (actually a named
list of \code{lazy} expressions).
}
\description{
Define the values characterising a Markov Model state for
1 cycle.
}
\details{
As with \code{\link[=define_parameters]{define_parameters()}}, state values are
defined sequentially. Later state definition can thus
only refer to values defined earlier.

For the \code{modify} function, existing values are
modified, no new values can be added. Values order
matters since only values defined earlier can be
referenced in later expressions.

If \code{count_types} is not specified, it defaults to "".
When values are calculated, this will be replaced
by the \code{method} argument of \code{\link[=eval_strategy]{eval_strategy()}}.
}
\examples{
st <- define_state(
  cost = 6453,
  utility = .876
)
st
}
