% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{define_state_transition_list}
\alias{define_state_transition_list}
\alias{define_state_transition_list_}
\title{Define Markov Model State Transition List}
\usage{
define_state_transition_list(...)

define_state_transition_list_(.dots)
}
\arguments{
\item{...}{expressions defining model state transtiions.}

\item{.dots}{List of states, only used by
\code{define_state_list_} to avoid using \code{...}.}

\item{.OBJECT}{An \code{uneval_state_transition_list} object.}
}
\value{
An object of class \code{uneval_state_list} (a
list of \code{state} objects).
}
\description{
Define the state transitions of a Markov model by combining
\code{state_transition} objects.
}
\details{
All states transitions should have the same value names
as model states.

The \code{modify} function can modify existing state
transitions or add new ones.
}
\examples{
\dontrun{
s1 <- define_state_transition(from = "A", cost = 1, util = 1)
s2 <- define_state_transition(from = "B", cost = 3, util = .4)

states_trans_mod <- define_state_transition_list(
  s1,
  s2
)

states_mod

s1_bis <- define_state_transition_list(from = "A", cost = 0, util = 1)
s3 <- define_state(to = "A", cost = 10, util = .1)

modify(
  states_mod,
  s1_bis,
  s3
)
}
  
}
\keyword{internal}
