% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{define_surv_dist}
\alias{define_surv_dist}
\title{Define a Survival Distribution}
\usage{
define_surv_dist(distribution = c("exp", "weibull"), ...)
}
\arguments{
\item{distribution}{A parametric survival distribution.}

\item{...}{Additional distribution parameters (see xxx).}
}
\value{
A \code{surv_dist} object.
}
\description{
Define a parametric survival distribution, to be used
by \code{\link{get_probs_from_surv}}.
}
\examples{

define_surv_dist(distribution = "exp", lambda = .5)

}

