% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_surv_table}
\alias{define_surv_table}
\alias{define_surv_table.data.frame}
\alias{define_surv_table.character}
\title{Define a survival distribution based on explicit survival probabilities}
\usage{
define_surv_table(x)

\method{define_surv_table}{data.frame}(x)

\method{define_surv_table}{character}(x)
}
\arguments{
\item{x}{a data frame with columns \code{time} and \code{survival}}
}
\value{
a \code{surv_table} object, which can be used with \code{\link[=compute_surv]{compute_surv()}}.
}
\description{
Define a survival distribution based on explicit survival probabilities
}
\examples{
 x <- data.frame(time = c(0, 1, 5, 10), survival = c(1, 0.9, 0.7, 0.5))
 define_surv_table(x)
 
}
