% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_survival}
\alias{define_survival}
\title{Define a Survival Distribution}
\usage{
define_survival(distribution = c("exp", "weibull", "weibullPH", "lnorm",
  "llogis", "gamma", "gompertz", "gengamma", "gengamma.orig", "genf",
  "genf.orig"), ...)
}
\arguments{
\item{distribution}{A parametric survival distribution.}

\item{...}{Additional distribution parameters (see
respective distribution help pages).}
}
\value{
A \code{surv_dist} object.
}
\description{
Define a parametric survival distribution.
}
\examples{

define_survival(distribution = "exp", rate = .5)
define_survival(distribution = "gompertz", rate = .5, shape = 1)

}
