% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_define.R
\name{define_wtd_surv_table}
\alias{define_wtd_surv_table}
\alias{define_wtd_surv_table.data.frame}
\title{Title}
\usage{
define_wtd_surv_table(x, ...)

\method{define_wtd_surv_table}{data.frame}(x, time_col, weights, by_age)
}
\arguments{
\item{x}{a data frame, or a string pointing to such a data frame}

\item{time_col}{the column of x measuring time}

\item{weights}{a named vector of weights.   Names correspond to
the columns of \code{x} to which the weights should be applied.}

\item{by_age}{is time in the table measured by age instead of cycle
(as is typical for a mortality table, for example).}
}
\value{
a \code{surv_pooled} object (see \code{\link[=mix]{mix()}}).
}
\description{
Title
}
\details{
\code{weights} are not checked for adding to 1.
}
