% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_strategy.R
\name{dispatch_strategy}
\alias{dispatch_strategy}
\title{Dispatch Values According to Strategy}
\usage{
dispatch_strategy(.strategy, ...)
}
\arguments{
\item{.strategy}{Optional strategy name. If not specified
it is implicitely added.}

\item{...}{Values of the parameter named depending on the
strategy.}
}
\value{
A vector of values.
}
\description{
Returns different values depending on the strategy.
}
\examples{

define_parameters(
  val = 456,
  x = dispatch_strategy(
    strat_1 = 1234,
    strat_2 = 9876,
    strat_3 = val * 2 + markov_cycle
  )
)
}

