% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch_strategy.R
\name{dispatch_strategy_hack}
\alias{dispatch_strategy_hack}
\title{Hack to Automate Use of Strategy Name}
\usage{
dispatch_strategy_hack(.dots)
}
\arguments{
\item{.dots}{A \code{lazy_dots} object.}
}
\value{
A modified \code{lazy_dots} object.
}
\description{
This function is a hack to automate the definition of the
argument \code{.strategy} in
\code{\link{dispatch_strategy}}.
}
\details{
The hack consists in replacing calls to 
\code{dispatch_strategy(...)} by 
\code{dispatch_strategy(.strategy = strategy, ...)} if
\code{.strategy_name} is not already defined.
}
\keyword{internal}

