% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_eval.R
\name{eval_matrix}
\alias{eval_matrix}
\title{Evaluate Markov Model Transition Matrix}
\usage{
eval_matrix(x, parameters)
}
\arguments{
\item{x}{an \code{uneval_matrix} object.}

\item{parameters}{an \code{eval_parameters} object.}
}
\value{
An \code{eval_matrix} object (actually a list of 
  transition matrix, one per cycle).
}
\description{
Evaluate a transition matrix using evaluated parameters.
}
\details{
Runs checks on the transition matrix during evaluation.
}
\keyword{internal}

