% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_eval.R
\name{eval_parameters}
\alias{eval_parameters}
\title{Evaluate Markov model parameters}
\usage{
eval_parameters(x, cycles = 1, strategy_name = NA,
  max_state_time = cycles)
}
\arguments{
\item{x}{an \code{uneval_parameters} object.}

\item{cycles}{integer. Number of cycles to simulate.}
}
\value{
An object of class \code{eval_parameters}
(actually a data.frame with one column per parameter
and one row per cycle).
}
\description{
Evaluate parameters specified through
\code{define_parameters}, for a given number of cycles.
}
\examples{
param <- define_parameters(
  age_start = 60,
  age = age_start + markov_cycle
)

heemod:::eval_parameters(param, cycles = 15)
}
\keyword{internal}
