% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_eval.R
\name{eval_resample}
\alias{eval_resample}
\title{Evaluate Resampling Definition}
\usage{
eval_resample(psa, N, top_eval_env = eval_env(), top_caller_env = caller_env())
}
\arguments{
\item{psa}{A \code{\link[=define_psa]{define_psa()}} object.}

\item{N}{> 0. Number of simulation to run.}
}
\value{
A \code{data.frame} of resampled values with on
column per parameter and \code{N} rows.
}
\description{
Evaluate Resampling Definition
}
\keyword{internal}
