% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_eval.R
\name{eval_state_list}
\alias{eval_state_list}
\title{Evaluate Markov Model States}
\usage{
eval_state_list(x, parameters)
}
\arguments{
\item{x}{An \code{uneval_state_list} object generated by 
\code{\link{define_state_list}}.}

\item{parameters}{An \code{eval_parameters} object 
generated by \code{\link{eval_parameters}}.}
}
\value{
An \code{eval_states} object, a list with one 
  data.frame per state containing a column per state 
  value and a line per cycle.
}
\description{
Evaluate Markov Model States
}
\keyword{internal}

