% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_from_files.R
\name{execute_models}
\alias{execute_models}
\title{Execute models}
\usage{
execute_models(inputs, base_dir, N.prob = 50, cost, effect, init = NULL,
  base_model = NULL, run_demo = TRUE, run_psa = TRUE, ...)
}
\arguments{
\item{inputs}{a list containing the unevaluated models
($models),    parameters for  the base model and
sensitivity analyses ($param_info), and possibly an
environment containing additional information necessary
to run the model.}

\item{base_dir}{directory in which \code{ref-file} will
be found.}

\item{N.prob, cost, effect, init}{arguments to 
\code{\link{run_models}}}

\item{base_model}{which of the models should be the base
model}

\item{run_demo, run_psa}{Useful for limiting computation,
for example, when running a model interactively in
shiny.}

\item{...}{additional arguments}
}
\value{
an unevaluated models object

a list \itemize{ \item \code{model}s (always) 
  \item \code{model_runs} (always) \item \code{dsa}
  (deterministic sensitivity analysis) - if appropriate
  parameters provided \item \code{psa} (probabilistic
  sensitivity analysis) - if appropriate parameters
  provided \item \code{demographics} results across
  different demographic groups }
}
\description{
Execute a full set of analyses, possibly
  including discrete sensitivity analysis, probabilistic
  sensitivity analysis, and analyses across demographics.
}
\details{
Relies on the inputs collated by 
\code{\link{gather_model_info}}. These pieces were broken
apart to facilitate putting together a shiny interface.

For more information see the vignette: 
\code{vignette("file-input", package = "heemod")}.`
}

