% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_state}
\alias{expand_state}
\alias{expand_state.uneval_matrix}
\alias{expand_state.uneval_state_list}
\title{Expand Time-Dependant States into Tunnel States}
\usage{
expand_state(x, ...)

\method{expand_state}{uneval_matrix}(x, state_pos, state_name, cycles, n = 1)

\method{expand_state}{uneval_state_list}(x, state_name, cycles)
}
\arguments{
\item{x}{A transition matrix or a state list.}

\item{...}{Addition parameters passed to methods.}

\item{state_pos}{Position of the state to expand.}

\item{state_name}{Original name of the sate to expand.}

\item{cycles}{Number of cycle of the model.}

\item{n}{Postition in the expansion process.}
}
\value{
The same object type as the input.
}
\description{
This function for transition matrices and state values 
expands states relying on \code{state_cycle} in a serie
of tunnels states.
}
\keyword{internal}

