% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi.R
\name{export_savi}
\alias{export_savi}
\title{Export PSA Results for SAVI}
\usage{
export_savi(x, folder = "~")
}
\arguments{
\item{x}{PSA result.}

\item{folder}{A folder where to save the \code{csv} files.}
}
\value{
Nothing. Creates 3 files.
}
\description{
Export the result of a PSA in a format compatible with 
Sheffield Accelerated Value of Information sofware.
}
\details{
This function saves 3 files at the path given by
\code{folder}: \code{param.csv}, the parameter values,
\code{cost.csv} and \code{effect.csv} the cost and effect
results.

The official SAVI website can be found at this URL: 
http://http://savi.shef.ac.uk/SAVI/
}

