% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{extract_stratum}
\alias{extract_stratum}
\title{Extract Product-Limit Table for a Stratum}
\usage{
extract_stratum(sf, index)
}
\arguments{
\item{sf}{A survfit object.}

\item{index}{The index number of the strata to extract.}
}
\value{
A data frame of the product-limit table for the
given stratum.
}
\description{
Extracts the product-limit table from a survfit object
for a given stratum. Only \code{\link[survival:survfit]{survival::survfit()}} objects are
supported.
}
\keyword{internal}
