% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{extract_surv_fit_metrics}
\alias{extract_surv_fit_metrics}
\title{Title}
\usage{
extract_surv_fit_metrics(fit_tib, metric = c("AIC", "BIC", "m2LL"))
}
\arguments{
\item{fit_tib}{a tibble of fits, for example from
\code{\link[=survival_fits_from_tabular]{survival_fits_from_tabular()}}}

\item{metric}{which metrics to add}
}
\value{
a tibble with the added columns
}
\description{
Title
}
