% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{extractParams}
\alias{extractParams}
\title{Extract Evaluated Parameters}
\usage{
extractParams(obj, data = NULL)
}
\arguments{
\item{obj}{A flexsurv object.}

\item{data}{An optional dataset of covariate values to generate parameters
for.  Defaults to the original data to which the model was fit.}
}
\value{
A tidy data frame of curve parameters for each covariate level.
}
\description{
Extracts the covariate-adjusted parameters from a flexsurv object.
}
\keyword{internal}
