% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{extractStrata}
\alias{extractStrata}
\title{Extract Product-Limit Tables}
\usage{
extractStrata(sf)
}
\arguments{
\item{sf}{A survit object.}
}
\value{
A tidy data.frame of the product-limit tables for all strata.
}
\description{
Extracts the product-limit table from a survfit object for all strata.
Only \code{survfit} and unstratified \code{survfit.coxph} objects are supported.
}
\keyword{internal}
