% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{extractStratum}
\alias{extractStratum}
\title{Extract Product-Limit Table for a Stratum}
\usage{
extractStratum(sf, index)
}
\arguments{
\item{sf}{A survit object.}

\item{index}{The index number of the strata to extract.}
}
\value{
A data.frame of the product-limit table for the given stratum.
}
\description{
Extracts the product-limit table from a survfit object for a given stratum.
Only \code{survfit} and unstratified \code{survfit.coxph} objects are supported.
}
\keyword{internal}
