% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_create_model_from_tabular}
\alias{f_create_model_from_tabular}
\title{Create a heemod model based on information in csv files.}
\usage{
f_create_model_from_tabular(state_file, tm_file)
}
\arguments{
\item{state_file}{The name of a csv file to be used by 
\code{f.create.state.definitions.from.tabular}.}

\item{tm_file}{The name of a csv file to be used by 
\code{f.transition.prob.matrix.from.tabular}.}
}
\value{
A heemod model as returned by \code{define_model}.
}
\description{
Create a heemod model based on information in csv files.
}
\details{
For more information see the vignette: 
     \code{vignette("tabular-input", package = "heemod")}.`
}

