% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_create_state_definitions_from_tabular}
\alias{f_create_state_definitions_from_tabular}
\title{Create a state definition from tabular input.}
\usage{
f_create_state_definitions_from_tabular(state_info)
}
\arguments{
\item{state_info}{A data frame or the name of a file containing the information.}
}
\value{
A list with three elements:
  \itemize{
   \item state_def_string The string that would get passed
   to \code{\link{define_state_list}}. 
   \item state_command The The call to create the states - 
   \code{\link{define_state_list}} with the previous item.
   \item state_names The names of the resulting states,
     for convenience.}
}
\description{
Takes tabular input defining states for 
  multiple models and creates a string to be passed to
  \code{\link{define_model}}.
}
\details{
Columns of state_info besides .model and state include 
     costs and utilities we want to keep track of,
     with appropriate values (these may include parameters).   
     For any cost
     or utility that should be discounted, 
     an additional column with the name
     ".discount.<cost>" or ".discount.<effect>", 
     for the appropriate cost or effect, can be included.
     If no discounting is desired for a particular
     cost or effect, the corresponding column can be omitted.   
     
     A discount column
     can contain only a single value - a cost or benefit
     must be discounted by the same amount in each state.
     Discounts can be numbers or parameters (which will then need to be defined
     like any other).

     The input data frame is expected to contain state information
     for all the models you will use in an analysis.
     For more information see the vignette: 
     \code{vignette("file-input", package = "heemod")}.`
}

