% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{f_fit_survival_models}
\alias{f_fit_survival_models}
\title{Fit survival models using different distributions}
\usage{
f_fit_survival_models(survdata, dists = allowed_fit_distributions,
  time_col_name, censor_col_name, treatment_col_name,
  covariate_col_names = NULL, fit_indiv_groups = TRUE)
}
\arguments{
\item{survdata}{Survival data to be used.}

\item{dists}{Distributional forms to be considered in fitting using
\code{flexsurvreg}.  By default, includes exponential, Weibull,
lognormal, log-logistic, gamma, gompertz, and generalized gamma.  Kaplan-Meier
curves will also be stored automatically, with distribution
name "km".}

\item{time_col_name}{Name of the column in survdata with event times.}

\item{censor_col_name}{Name of the column in survdata with censorship
indicators.}

\item{treatment_col_name}{Name of the column in survdata that holds the
treatment group names (for example "control", "treatment",
"dose1", and so on).}

\item{covariate_col_names}{Not yet implemented}

\item{fit_indiv_groups}{Should groups be fit individually?  Default TRUE.}
}
\value{
a matrix (with dimnames) of flexsurvreg objects,
with rows corresponding to
the distributions, and columns corresponding to groups.
}
\description{
Fit survival models using different distributions
}
\details{
\code{survdata} should be a data frame of the form required for
\code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}}.   In the column specified
by \code{censor_col_name} in the data itself,
1 means that an event (frequently death or disease progression,
depending on context) was observed, while 0
means the event was censored.  So the Kaplan-Meier plot will have
a drop anywhere the censor column contains a 1, and will not
contain a drop when the censor column contains a 0.
}
