% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{f_get_best_surv_model}
\alias{f_get_best_surv_model}
\title{Choose the best model out of a set based on a metric.}
\usage{
f_get_best_surv_model(surv_fits, metric = c("AIC"))
}
\arguments{
\item{surv_fits}{A list object from \code{f.get.survival.probs} that gives
a collection (list) of parametric survival fit object.}

\item{metric}{The metric to choose the model by.
Currently supports selecting the best model using one (and only one) of the
following metrics:
1. "AIC": Akaike information criterion
2. "BIC": Bayesian information criterion
3. "m2LL": -2*log likelihood}
}
\value{
The best model according to the chosen metric.
}
\description{
Choose the best model out of a set based on a metric.
}
\details{
Gets the best survival model, according to a particular metric.
Current implementation is limited to selecting the best model
using one (and only one) of the following metrics:
1. Akaike information criterion (AIC)
2. Bayesian information criterion (BIC)
3. -2*log likelihood (-2LL)
}
