% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_look_up_values_df.R
\name{f_look_up_values_df}
\alias{f_look_up_values_df}
\title{Look up values in a data frame}
\usage{
f_look_up_values_df(val_source, ..., numeric_cols = NULL, output_col = NULL)
}
\arguments{
\item{val_source}{A data frame with attributes having various indices
and a value (or several values) to be looked up}

\item{...}{Collection of indices to look-up for each attribute.}

\item{numeric_cols}{names specifying which attributes should be treated as numeric.}

\item{output_col}{If specified, the name of the column to get for output.}
}
\value{
A single value or vector of values (depending on how many attribute sets
		were provided).
}
\description{
Read in values from a data frame, typically one imported
  using \code{\link{load_df_from_files}}.	For example, could be used
  for analysis-specific mortality rates.
}
\details{
Columns designated as numeric by \code{numeric_cols} get looked up using
   ranges of values rather than by single values. In this case, each attribute index
		is the minimum value for a bin that ends with the next index.
		
		The order of the \code{...} arguments need not match the order 
		of columns in the data frame. 
 		
		If specified, \code{output_col} must be the name of a single column to get for 
		output.   If not specified, it
	  will be set to the difference between names(val_source) and the names of index
	  columns, and an error will be thrown if that difference contains more than one column.
}
\examples{
 tempdf <- expand.grid(arg1 = c("A", "B", "C"), arg2 = 1:4, arg3 = 1:5)
 tempdf$value <- 1:60
 
 f_look_up_values_df(tempdf, 
                     arg1 = c("A", "B", "C", "B", "A"),
                     arg2 = c(1, 1, 3.2, 3.0, 5), 
                     arg3 = c(1, 1, 1, 2, 3),
                     numeric_cols  = c("arg2", "arg3"))
 
 age_related_df <- data.frame(age = 10 * 0:9, decade = 1:10)
 
 f_look_up_values_df(age_related_df, age = c(0, 10, 20), output_col = "decade")
 f_look_up_values_df(age_related_df, age = c(5, 15, 23.5), 
                     numeric_cols = "age", output_col = "decade")

## error if you don't specify that a numeric column is numeric
##   and you don't have values that are in the table
\dontrun{
 f_look_up_values_df(age_related_df, age = c(5, 15, 23.5), output_col = "decade")
 }
}

