% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_parameters_from_tabular}
\alias{f_parameters_from_tabular}
\title{Create a heemod parameter file from tabular input.}
\usage{
f_parameters_from_tabular(input_file_name, output_file_name_base = NULL,
  param_obj_name = "params")
}
\arguments{
\item{input_file_name}{character. Contains the name (and path) to 
R, CSV, XLS or XLSX file with parameters.}

\item{output_file_name_base}{character. If not NULL, the parameters will be
saved as .R files with this prefix. NULL by default.}

\item{param_obj_name}{base name for parameter objects}
}
\value{
the parameter definition, invisibly.
}
\description{
Create a heemod parameter file from tabular input.
}
\details{
The tabular parameter definition file can have the following columns:
  parameter (the parameter name, required), value (required), 
  low and high (if both are present, a deterministic sensitivity analysis 
  will be performed), and psa (a definition of a distribution to use in a
  probabilistic sensitivity analysis.
  Other columns can exist, but will be ignored.
  
  If it ends in ".csv", ".xls", or ".xlsx",
  name of a .csv file defining the parameters (see details for contents).  
  If the file name
  ends in ".R", a parameter definition file that can
  be sourced.  (This can be useful if you've previously read
  a  parameter definition and want to make modifications
  directly in the R.)
  
  If output_file_name_base is not \code{NULL} (the default), then it is the
  file name into which a heemod parameter definition
  file will be written.  The basic parameter file will get extension ".R",
  and the discrete and probabilistic sensitivity analysis parameter 
  definition files (if 
  created) will get extensions ".dsa.R" and ".psa.R" respectively.   
  If \code{NULL}, no file will be written.
  
     For more information see the vignette: 
     \code{vignette("file-input", package = "heemod")}.`
}

