% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_parse_multi_spec}
\alias{f_parse_multi_spec}
\title{Specify inputs for multiple models in a single file}
\usage{
f_parse_multi_spec(multi_spec, split_on = ".model", group_vars)
}
\arguments{
\item{multi_spec}{\code{data frame}, or, if of type 
\code{character}, the name of a file that contains the data frame.}

\item{split_on}{\code{character}, with the name of the variable
in multi_spec to be split on.}

\item{group_vars}{\code{character}, one or more variable names 
from \code{multi_spec}.}
}
\value{
a list of data frames, one for each value of split_on.
}
\description{
Parse a matrix containing specifications for 
multiple models into the inputs required for each model.
}
\details{
Each combination of values of the columns specified by 
 \code{group_vars} should either be unique in the file (in which
  case it will be replicated for all values of split_on), 
  or must be repeated as many times as unique values of split_on.

     For more information see the vignette: 
     \code{vignette("file-input", package = "heemod")}.`
}

