% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_read_demographic_table}
\alias{f_read_demographic_table}
\title{Read in demographic table.}
\usage{
f_read_demographic_table(demographicTable, weight_col_name = "weight", params,
  ...)
}
\arguments{
\item{demographicTable}{A demographic table, which contains one column for each
stratifying variable in the model (for example, age and sex).}

\item{weight_col_name}{The name of the column that contains the weight of each
demographic group defined by the stratifying variables.}

\item{params}{Parameters of a model, to check that all the columns in the demographic
table (other than the weight column) are in the model.}

\item{...}{catches other, unwanted arguments.}
}
\value{
A data frame with the information from the file.
}
\description{
Read in demographic table.
}
\details{
The demographic table should be a data frame with n+1 columns, where there are
  n variables we use to stratify the population.    The first n columns
  correspond to the stratification variables (for example sex and age),
  and the column names must correspond to the variables used in calculating
  probabilities in the parameters object.
  The n+1-st column gives the proportion of the population in each group.
  The values in the n+1-st column should add to 1; if they do not, they will
  be renormalized, with a warning.
}

