% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_read_file}
\alias{f_read_file}
\title{Read the accepted file formats for tabular input}
\usage{
f_read_file(file_name)
}
\arguments{
\item{file_name}{name of the file}
}
\value{
The file, as a data frame, with strings not converted to factors.
Columns containing 'comment' in the header (case in-sensitive) are ignored. #'
}
\description{
Read the accepted file formats for tabular input
}

