% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{f_transition_prob_matrix_from_tabular}
\alias{f_transition_prob_matrix_from_tabular}
\title{Create a transition matrix from tabular input.}
\usage{
f_transition_prob_matrix_from_tabular(trans_probs, state_names)
}
\arguments{
\item{trans_probs}{A data frame or the name of a data frame.}

\item{state_names}{The names of the states used in the transition matrix.}
}
\value{
A list with two elements:  
   \itemize{
   \item \code{string}, a character string that can be used 
   in defining a heemod model in 
   \code{\link{f_create_model_from_tabular}}
   \item
   \code{transition_matrix}, the matrix itself 
   (created by evaluating the command using 
   \code{\link{define_matrix}} exactly
   as though the text had been entered at the command line).
   }
}
\description{
Create a transition matrix from tabular input.
}
\details{
The data frame \code{trans_probs} should have columns
 \code{from}, \code{to}, and \code{prob}, where \code{prob} 
 is the probability
  of a transition from the \code{from} state to the \code{to} state.  
  Prob can be defined in terms of parameters, 
  just as when using \code{define_matrix} at the keyboard.
  Probabilities of 0 need not be specified - 
  they will be automatically inserted.
  
  All state names must be used in the "from" column of the 
  transition matrix (otherwise you can just get rid of the state). 
  Absorbing states should have
  a transition from and to themselves with probability 1.
   
  If trans_probs is a string, it is the file name of a file
  (.csv, .xls, or .xlsx) containing the data frame.
     The input data frame is expected to contain transition 
     probabilities for all the models you will use in an analysis.
     For more information see the vignette: 
     \code{vignette("file-input", package = "heemod")}.`
}

