% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{is_csv}
\alias{is_csv}
\alias{is_xls}
\alias{is_xlsx}
\title{Check File Type}
\usage{
is_csv(x)

is_xlsx(x)

is_xls(x)
}
\arguments{
\item{x}{A file name.}
}
\value{
Whether the file is (respectively)
 csv, xlsx, or xls.
}
\description{
Check File Type
}
\keyword{internal}

