% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{is.csv}
\alias{is.csv}
\alias{is.xls}
\alias{is.xlsx}
\title{File checkers}
\usage{
is.csv(input_file_name)

is.xlsx(input_file_name)

is.xls(input_file_name)
}
\arguments{
\item{input_file_name}{The file name}
}
\value{
whether the file is (respectively) csv, xlsx, or xls
}
\description{
File checkers
}

