% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{filter_blanks}
\alias{filter_blanks}
\title{Remove Blank Rows From Table}
\usage{
filter_blanks(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\value{
A \code{data.frame} without blank rows.
}
\description{
Remove rows were all values are \code{NA}.
}
\details{
Some rows can be left blanks in the input table for 
readability, this function ensures those rows are 
removed.
}
\keyword{internal}

