% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{fit_survival}
\alias{fit_survival}
\title{Fit Survival Models}
\usage{
fit_survival(x, dists = c("exp", "weibull", "lnorm", "gamma", "gompertz",
  "gengamma"), time, censor, group = NULL, covariates = NULL,
  fit_by_group = TRUE)
}
\arguments{
\item{x}{Survival data to be used.}

\item{dists}{Distributional forms to be considered in 
fitting using \code{flexsurvreg}.}

\item{time}{Event times.}

\item{censor}{Censorship indicators. 0: event 
observed; 1: censored.}

\item{group}{Optional treatment group names.}

\item{covariates}{Not yet implemented.}

\item{fit_by_group}{Should groups be also fit 
individually?}
}
\value{
a matrix of flexsurvreg objects, 
  with rows corresponding to distributions, and 
  columns corresponding to groups.
}
\description{
Fit multiple survival models, with various distributions.
}
\details{
No objects are returned for the Kaplan-Meier fit; that 
can be obtained from any of the fitted objects by 
using the \code{summary.flexsurvreg} function with
\code{type = "survival"}.
}

