% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_best_fit}
\alias{get_best_fit}
\title{Return Best Fitting Distribution}
\usage{
get_best_fit(x, metric = c("AIC", "BIC", "m2LL"))
}
\arguments{
\item{x}{Object returned by \code{\link{fit_survival}}.}

\item{metric}{Metric to assess model fit.}
}
\value{
The best fitting distribution according to the
  chosen metric.
}
\description{
Given an object from \code{\link{fit_survival}}, returns 
the best fitting distribution with regard to the given
metric.
}
\details{
Gets the best survival fit, according to a particular 
metric. Current metrics are: Akaike information criterion
(\code{AIC}), Bayesian information criterion 
(\code{BIC}), -2*log(likelihood) (\code{m2LL}).
}

