% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\alias{get_code.eval_model_list}
\alias{get_code.state}
\alias{get_code.uneval_matrix}
\alias{get_code.uneval_model}
\alias{get_code.uneval_parameters}
\alias{get_code.uneval_state_list}
\title{Display the Code to Generate an Object}
\usage{
get_code(x, ...)

\method{get_code}{uneval_parameters}(x, name = NULL, ...)

\method{get_code}{uneval_matrix}(x, name = NULL, ...)

\method{get_code}{state}(x, name = NULL, ...)

\method{get_code}{uneval_state_list}(x, ...)

\method{get_code}{uneval_model}(x, name = NULL, sub = FALSE, ...)

\method{get_code}{eval_model_list}(x, name = NULL, sub = FALSE, ...)
}
\arguments{
\item{x}{An \code{heemod} object.}

\item{...}{Additional parameters passed to methods.}

\item{name}{character. Optional argument giving the name to assign to the object.}

\item{sub}{logical. Should states or models be referenced by name in \code{define_model}
and \code{run_model} instead of including the entire code?}
}
\value{
A character string.
}
\description{
This function returns the \code{R} code to generate an \code{heemod} oj
}
\examples{

p <- define_parameters(
  a = 2,
  b = 3 * x + 1
)

get_code(p)

m <- define_matrix(
  C, .1,
  0, 1 
)

get_code(m)

}

