% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_define.R
\name{get_counts}
\alias{get_counts}
\title{Get State Membership Counts}
\usage{
get_counts(x, ...)
}
\arguments{
\item{x}{Result from \code{\link{run_models}}.}

\item{m}{Model name or index.}
}
\value{
A data frame of counts per state.
}
\description{
Given a result from \code{\link{run_models}}, return
state membership counts for a specifiec model.
}

