% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_counts_diff}
\alias{get_counts_diff}
\title{Get count matrix and difference between two cycles}
\usage{
get_counts_diff(x, init, inflow)
}
\arguments{
\item{x}{transition matrix}

\item{init}{numeric vector, same length as number of
model states. Number of individuals in each model state
at the beginning.}

\item{inflow}{numeric vector, similar to \code{init}.
Number of new individuals in each state per cycle.}
}
\value{
A length 2 list of matrix : the count matrix for each cycle and the diff matrix
showing the difference of counts between two cycles.
}
\description{
Get count matrix and difference between two cycles
}
\keyword{internal}
