% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_define.R
\name{get_init}
\alias{get_init}
\title{Get Initial State Values}
\usage{
get_init(x)
}
\arguments{
\item{x}{x Result from \code{\link{run_models}}.}
}
\value{
A vector of initial state values.
}
\description{
Get Initial State Values
}

