% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mat_total}
\alias{get_mat_total}
\title{Get count number}
\usage{
get_mat_total(x, init)
}
\arguments{
\item{x}{transition matrix}

\item{init}{numeric vector, same length as number of
model states. Number of individuals in each model state
at the beginning.}
}
\value{
A count matrix
}
\description{
Get count number
}
\keyword{internal}
