% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_define.R
\name{get_matrix_order}
\alias{get_matrix_order}
\title{Return Markov Model Transition Matrix Order}
\usage{
get_matrix_order(x)
}
\arguments{
\item{x}{A transition matrix, evaluated or not.}
}
\value{
An integer: matrix order.
}
\description{
A generic that works both with \code{uneval_matrix} and
\code{eval_matrix}.
}
\details{
For internal use.
}
\keyword{internal}

