% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model}
\alias{get_model}
\title{Get values from a specific model}
\usage{
get_model(x, m)
}
\arguments{
\item{x}{A data.frame with results from several models.}

\item{m}{Model name or position.}
}
\value{
A data.frame with results from only one model.
}
\description{
Get values from a specific model
}

