% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mortality_rate}
\alias{get_mortality_rate}
\title{Get Mortality Rate}
\usage{
get_mortality_rate(age, sex, ref)
}
\arguments{
\item{age}{Age.}

\item{sex}{Sex.}

\item{ref}{Reference mortality rate table, see details
for structure.}
}
\value{
A vector of mortality rates.
}
\description{
Convenience function for non-homogeneous models. Given 
age, sex and a reference table, returns mortality rate.
}
\details{
The \code{ref} table must contain 3 columns named 
\code{age}, \code{sex} and \code{rate}.
}
\examples{

age_seq <- 31:38
sex_seq <- rep(1:2, 4)

tab_ref <- data.frame(
  age = c("[30,35)", "[30,35)", "[35,40)", "[35,40)"),
  sex = c(1, 2, 1, 2),
  rate = c(.02, .01, .05, .04)
)

get_mortality_rate(
  age = cut(age_seq, c(30, 35, 40), right = FALSE),
  sex = sex_seq,
  ref = tab_ref
)

}

