% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Return parameters names}
\usage{
get_parameter_names(x)
}
\arguments{
\item{x}{An object with parameters.}
}
\value{
A character vector of parameter names.
}
\description{
Extract parameters names.
}
\keyword{internal}
