% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Markov Model Parameters}
\usage{
get_parameters(x)
}
\arguments{
\item{x}{An \code{eval_model_list}
object.}
}
\value{
An \code{uneval_parameters} or
  \code{eval_parameters} object.
}
\description{
For internal use.
}

