% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_probs_from_surv}
\alias{compute_counts_part_surv}
\alias{get_probs_from_surv}
\alias{get_probs_from_surv.flexsurvreg}
\alias{get_probs_from_surv.surv_dist}
\title{Extract Transition Probabilities from a Survival Model}
\usage{
get_probs_from_surv(x, ...)

\method{get_probs_from_surv}{flexsurvreg}(x, cycle, km_limit = 0,
  cycle_length = 1, type = c("prob", "surv"), ...)

\method{get_probs_from_surv}{surv_dist}(x, cycle, km_limit = 0,
  cycle_length = 1, type = c("prob", "surv"), ...)

compute_counts_part_surv(part_surv_obj, use_km_until, num_patients,
  markov_cycle, markov_cycle_length, state_names)
}
\arguments{
\item{x}{Either a result from 
\code{\link[flexsurv]{flexsurvreg}} or
\code{\link{define_survival}}.}

\item{...}{arguments passed to methods.}

\item{cycle}{The \code{markov_cycle} or 
\code{state_cycle} for which to predict.}

\item{km_limit}{Up to what time should Kaplan-Meier 
estimates be used? Model predictions will be used 
thereafter. See \code{Details}.}

\item{cycle_length}{The value of a Markov cycle in 
absolute time units.}

\item{type}{either \code{prob}, for transition 
probabilities, or \code{surv}, for survival}

\item{part_surv_obj}{A \code{flexsurvreg} object, or a list
specifying a distribution.}

\item{use_km_until}{to what time should Kaplan-Meier estimates be used?
Model predictions will be used thereafter.}

\item{num_patients}{The number of patients being modeled.  Survival
probabilities are between 0 and 1, so we multiply up by num_patients.}

\item{state_names}{Name of the states to be assigned to the counts.}
}
\value{
Returns the Markov transition probability for the
  cycles.

A \code{cycle_counts} object.
}
\description{
Get probabilities from survival models.

Given a partitioned survival model, returns the number of individuals 
per state per cycle.
}
\details{
If \code{use_km_until = 0}, then only model probabilities
will be used.

See \code{get_probs_from_surv} for further information on
  \code{part_surv_obj}, \code{use_km_until}, and the \code{markov_cycle}
  arguments.
}
\keyword{internal}

