% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{get_state_names}
\alias{get_state_names}
\title{Get State Names}
\usage{
get_state_names(x, ...)
}
\arguments{
\item{x}{An object containing states.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of state names.
}
\description{
Retrieve state names from an object containing states.
}
\keyword{internal}

