% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{get_state_number}
\alias{get_state_number}
\title{Return Number of State}
\usage{
get_state_number(x)
}
\arguments{
\item{x}{An object containing states.}
}
\value{
An integer: number of states.
}
\description{
For internal use.
}
\details{
Work with both \code{uneval_states} and
\code{eval_states}.
}
\keyword{internal}

