% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{get_state_value_count_types}
\alias{get_state_value_count_types}
\title{Get value count types}
\usage{
get_state_value_count_types(x)
}
\arguments{
\item{x}{A \code{state} or \code{lazy_dot} or \code{uneval_state_list} object.}
}
\value{
a named vector of count types
}
\description{
Get value count types
}
\details{
The value count types will be passed to
\code{\link[=correct_counts]{correct_counts()}} as the \code{method} argument.
}
