% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states_define.R
\name{get_state_value_names}
\alias{get_state_value_names}
\title{Return Names of State Values}
\usage{
get_state_value_names(x)
}
\arguments{
\item{x}{An object containing states.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of state value names.
}
\description{
Return Names of State Values
}
\keyword{internal}

