% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_surv_probs}
\alias{get_surv_probs}
\title{Get transition or survival probabilities from fits}
\usage{
get_surv_probs(fits, treatment, km_until, markov_cycle, markov_cycle_length,
  pred_type = c("prob", "surv"))
}
\arguments{
\item{fits}{Fits from \code{flexsurvreg}, or a list specifying 
distributions}

\item{treatment}{The treatment for which probabilities are desired}

\item{km_until}{Up to what time should Kaplan-Meier estimates be used?}

\item{markov_cycle}{The Markov cycles for which to predict.}

\item{markov_cycle_length}{The value of a Markov cycle in absolute time units.}

\item{pred_type}{"prob" or "surv", depending on which type of prediction
is desired.}
}
\value{
a vector of probabilities
}
\description{
Get transition or survival probabilities from fits
}

