% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{get_survival_input}
\alias{get_survival_input}
\title{Make sure survival inputs have the correct formats}
\usage{
get_survival_input(ref)
}
\arguments{
\item{ref}{- data frame from the reference file}
}
\value{
a list with elements \itemize{\item{fit_files},
  \item{fit_names}, \item{surv_data_files}, \item{fit_metric}}.
}
\description{
Make sure survival inputs have the correct formats
}
\details{
For survival analysis, we need, for each condition
  (frequently progression-free survival and overall survival)
  at least two, and possibly three, elements represented in
  the reference file.
  \itemize{
    \item{\code{surv_data_file}:  the name of the file 
       with the survival data}
    \item{\code{fit_file}: the name of a file with fits.  If
    \code{surv_data_file} is not provided, the fit will
    be read in from this file; if \code{surv_data_file}
    is provided, the fit will be written into this file.}
    \item{\code{fit_name}:  name to give the survival fit so
    that it can be referred to to extract probabilities.}
    } 
At least one of \code{surv_data_file} and \code{fit_file}
  must be provided.
}

