% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_define.R
\name{get_transition}
\alias{get_transition}
\title{Get Markov Model Transition Matrix}
\usage{
get_transition(x)
}
\arguments{
\item{x}{An \code{uneval_model} or \code{eval_model} 
object.}
}
\value{
An \code{uneval_matrix} or \code{uneval_matrix} 
  object.
}
\description{
Works on both unevaluated and evaluated models.
}
\keyword{internal}

