% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_mortality.R
\name{get_who_mr_}
\alias{get_who_mr}
\alias{get_who_mr_}
\title{Use WHO Mortality Rate}
\usage{
get_who_mr_(age, sex, country, year = "latest")

get_who_mr(age, sex, country, year = "latest")
}
\arguments{
\item{age}{age as a continuous variable.}

\item{sex}{sex as \code{"FMLE"} or \code{"MLE"}.}

\item{country}{Country code (see details).}

\item{year}{Use data from that year. Defaults to \code{"latest"}.}
}
\value{
This function should be used within
 \code{\link{define_matrix}} or \code{\link{define_parameters}}.
}
\description{
Use WHO Mortality Rate
}
\examples{

define_matrix(
  C, get_who_mr(age = 50 + markov_cycle, sex = "FMLE", country = "FRA"),
  0, 1
)

}

