% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heemod.R
\docType{package}
\name{heemod-package}
\alias{heemod}
\alias{heemod-package}
\title{Markov Models for Health Economic Evaluations}
\description{
An implementation of the modelling and
reporting features described in reference
textbooks and guidelines: deterministic and
probabilistic sensitivity analysis,
heterogeneity analysis, time dependency
on state-time and model-time (semi-Markov
and non-homogeneous Markov models), etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aphp.github.io/heemod/}
  \item Report bugs at \url{https://github.com/aphp/heemod/issues}
}

}
\author{
\strong{Maintainer}: Kevin Zarca \email{kevin.zarca@gmail.com}

Authors:
\itemize{
  \item Antoine Filipovic-Pierucci
}

Other contributors:
\itemize{
  \item Matthew Wiener [contributor]
  \item Zdenek Kabat [contributor]
  \item Vojtech Filipec [contributor]
  \item Jordan Amdahl [contributor]
  \item Yonatan Carranza Alarcon [contributor]
  \item Vince Daniels [contributor]
}

}
\keyword{internal}
