% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heemod.R
\docType{package}
\name{heemod}
\alias{heemod}
\alias{heemod-package}
\title{Markov Models for Health Economic Evaluations}
\description{
An implementation of the modelling and
reporting features described in reference
textbooks and guidelines: deterministic and
probabilistic sensitivity analysis,
heterogeneity analysis, time dependency
on state-time and model-time (semi-Markov
and non-homogeneous Markov models), etc.
}
