% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R, R/resamp_print.R, R/run_model_summary.R, R/sensitivity_print.R
\name{scale.combined_model}
\alias{heemod_scale}
\alias{scale.combined_model}
\alias{scale.dsa}
\alias{scale.psa}
\alias{scale.run_model}
\title{Normalize Cost and Effect}
\usage{
\method{scale}{combined_model}(x, center = TRUE, scale = TRUE)

\method{scale}{psa}(x, center = TRUE, scale = TRUE)

\method{scale}{run_model}(x, center = TRUE, scale = TRUE)

\method{scale}{dsa}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{Result of \code{\link{run_model}} or 
\code{\link{run_psa}}.}

\item{center}{Center results around base model?}

\item{scale}{Scale results to individual values?}
}
\value{
Input with normalized \code{.cost} and 
  \code{.effect}, ordered by \code{.effect}.
}
\description{
Normalize cost and effect values taking base model as a 
reference.
}
\keyword{internal}

