% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tabular.R
\name{import}
\alias{import}
\title{Import Analysis from Tabular Data}
\usage{
import(path)
}
\arguments{
\item{path}{Path to a folder containing the model information.

Import an analysis saved in a folder as tabular data.

A saved analysis follows a very specific format. Use
\code{\link{create_import_template}} or \code{\link{export}}
to populate the folder.}
}
\value{
Same result as \code{\link{run_models}}.
}
\description{
Import Analysis from Tabular Data
}

