% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_from_files.R
\name{import_models_from_files}
\alias{import_models_from_files}
\title{Read in models specified by files}
\usage{
import_models_from_files(base_dir, ref_file, base_model)
}
\arguments{
\item{base_dir}{directory in which \code{ref-file} will be found}

\item{ref_file}{name of a file specifying where model inputs will be found.}

\item{base_model}{which of the models should be the base model}
}
\value{
an unevaluated models object
}
\description{
Read in models specified by files
}
\details{
The reference file should have two columns, 
  \code{data} and \code{file}.
  The \code{data} column must include entries
  \code{state}, \code{tm}, and \code{parameters}, with the corresponding
  entries in \code{file} naming the files
  containing information on states, the transition matrix,
  and parameters.  The parser  also
  looks for \code{demographics}, for a file containing demographic
  data.
  
     For more information see the vignette: 
     \code{vignette("file-input", package = "heemod")}.`
}

