% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{interp_heemod}
\alias{interp_heemod}
\alias{interp_heemod.default}
\alias{interp_heemod.part_surv}
\alias{interp_heemod.state}
\alias{interp_heemod.uneval_matrix}
\alias{interp_heemod.uneval_state_list}
\title{Interpolate Lazy Dots}
\usage{
interp_heemod(x, ...)

\method{interp_heemod}{default}(x, more = NULL, ...)

\method{interp_heemod}{uneval_matrix}(x, ...)

\method{interp_heemod}{state}(x, ...)

\method{interp_heemod}{part_surv}(x, ...)

\method{interp_heemod}{uneval_state_list}(x, ...)
}
\arguments{
\item{x}{A parameter, transition matrix or state list
object.}

\item{...}{Addition parameters passed to methods.}

\item{more}{A list of expressions.}
}
\value{
An interpolated lazy dots object.
}
\description{
Sequencially interpolates lazy dots, optionnaly using 
external references.
}
\details{
The interpolation is sequencial: the second dot is 
interpolated using the first, the third using the 
interpolated first two, and so on.
}
\keyword{internal}

