% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{interpolate}
\alias{interpolate}
\alias{interpolate.default}
\alias{interpolate.uneval_matrix}
\alias{interpolate.state}
\alias{interpolate.part_surv}
\alias{interpolate.uneval_state_list}
\title{Interpolate Lazy Dots}
\usage{
interpolate(x, ...)

\method{interpolate}{default}(x, more = NULL, ...)

\method{interpolate}{uneval_matrix}(x, ...)

\method{interpolate}{state}(x, ...)

\method{interpolate}{part_surv}(x, ...)

\method{interpolate}{uneval_state_list}(x, ...)
}
\arguments{
\item{x}{A parameter, transition matrix or state list
object.}

\item{...}{Addition parameters passed to methods.}

\item{more}{A list of expressions.}
}
\value{
An interpolated lazy dots object.
}
\description{
Sequentially interpolates lazy dots, optionally using
external references.
}
\details{
The interpolation is sequential: the second dot is
interpolated using the first, the third using the
interpolated first two, and so on.
}
\keyword{internal}
