% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check Wholenumbers}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric.}

\item{tol}{the smallest positive floating-point number x
such that 1 + x != 1.}
}
\value{
A logical scalar.
}
\description{
Check Wholenumbers
}
\keyword{internal}
