% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{least_cost}
\alias{least_cost}
\title{Get the least cost for the requested doses}
\usage{
least_cost(x)
}
\arguments{
\item{x}{a set of combinations from \code{\link{find_least_cost_partition}}.}
}
\value{
a data frame with columns \code{total} and \code{cost}.
}
\description{
Get the least cost for the requested doses
}
\examples{
## see the examples for find_least_cost_partition.
}
