% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{Model piecewise linear decline in vaccine efficacy
or other variables}
\usage{
linear_interpolation(marked_cycles, marked_values, init_value = 1,
  change_per_cycle, max_change_cycles, age_start_change_per_cycle = NULL,
  cycle_length = 1, cycles = NULL, age_at_cycle_0, min_val = -Inf)
}
\arguments{
\item{marked_cycles}{cycles at which effectiveness
fractions are specified.}

\item{marked_values}{effectiveness fractions
at the corresponding cycles.}

\item{init_value}{a point at (0, init_value) will be
added}

\item{change_per_cycle}{can be specified instead of
\code{marked_cycles} and \code{marked_times}.   If both
are specified, \code{marked_cycles} and \code{marked_values}
take precedence.}

\item{max_change_cycles}{used with change_per_cycle}

\item{age_start_change_per_cycle}{at what age should the decline in
efficacy start (if not specified, there's no threshold).}

\item{cycle_length}{used to calculate age from cycles if necessary}

\item{cycles}{The cycles for which values should be returned.
If \code{NULL}, the default, a function is returned.}

\item{age_at_cycle_0}{used in conjunction with
\code{age_start_change_per_cycle}.}

\item{min_val}{minimum value the interpolation can take on.}
}
\value{
a function that calculates the decline in efficacy
based on cycle and, if specified, patient age at the
start of a simulation.
}
\description{
Model piecewise linear decline in vaccine efficacy
or other variables
}
\details{
intended for use in parameters object
for \code{heemod} models.
}
\examples{
linear_interpolation(c(1, 10, 20), c(1, .5, .25))
linear_interpolation(change_per_cycle = -0.02, max_change_cycles = 50)
linear_interpolation(change_per_cycle = -0.02, max_change_cycles = 100,
   min_val = -0.4)
linear_interpolation(age_start_change_per_cycle = 65, age_at_cycle_0 = 60,
  change_per_cycle = -0.05, cycles = 1:15, max_change_cycles = 5)

}
