% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_all_same}
\alias{list_all_same}
\title{Check if All the Elements of a List Are the Same}
\usage{
list_all_same(x)
}
\arguments{
\item{x}{a list.}
}
\value{
A logical scalar.
}
\description{
Check if All the Elements of a List Are the Same
}
\keyword{internal}

