% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_df_from_files.R
\name{load_df_from_files}
\alias{load_df_from_files}
\title{Load Data From a Folder Into an Environment}
\usage{
load_df_from_files(df_dir, df_envir)
}
\arguments{
\item{df_envir}{An environment.}

\item{full_path}{A directory containing the files.}
}
\value{
The environment with the data frames.
}
\description{
Reads files containing data frames (in tabular format) 
from a directory, and loads them in an environment to be 
available during an analysis.  Typically used with 
\code{\link{f_look_up_values_df}}.
}
\details{
The files must be in .csv, .xls, or .xlsx format. A file 
my_df.csv (or my_df.xls, or my_df.xlsx) will be loaded as
a data frame my_df.
}

