% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{load_surv_models}
\alias{load_surv_models}
\title{Load a set of survival fits}
\usage{
load_surv_models(location, survival_specs)
}
\arguments{
\item{location}{base directory}

\item{survival_specs}{information about fits}
}
\value{
a list with the fits for each data file passed in
}
\description{
Load a set of survival fits
}
