% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_names}
\alias{make_names}
\title{Make Syntactically Valid Names}
\usage{
make_names(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
Compared to \code{\link{make.names}} this function also 
converts characters to lower case and replaces \code{.}
by \code{_}.
}
\keyword{internal}

