% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{maximize_hazards}
\alias{maximize_hazards}
\title{Get the Maximum Hazards}
\usage{
maximize_hazards(...)
}
\arguments{
\item{...}{Survival distributions to be used in the projection.}
}
\value{
A \code{surv_max_haz} object.
}
\description{
Get a survival distribution reflecting the maximum of the hazard rates
in each cycle between two or more survival distributions.
}
\examples{

gp_mortality = define_survival(distribution = "exp", rate = .001)
disease_mortality = define_survival(
  distribution = "weibull",
  shape = 0.5,
  scale = 50
)
combined_dist = maximize_hazards(disease_mortality, gp_mortality)
}
