% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_define.R, R/part_surv.R
\name{modify_transition}
\alias{modify_transition}
\alias{modify_transition.part_surv}
\title{Modify an unevaluated transition object}
\usage{
modify_transition(x, ...)

\method{modify_transition}{part_surv}(x, transition_options)
}
\arguments{
\item{x}{a transition object}

\item{...}{additional arguments for modification}

\item{transition_options}{a list of options for modifying the transition object}
}
\value{
a modified version of the transition object
}
\description{
Modify an unevaluated transition object
}
\examples{
dist_pfs <- define_survival("exp", rate = 1)
dist_os <- define_survival("exp", rate = .5)

ps1 <- define_part_surv(
  pfs = dist_pfs,
  os = dist_os,
  terminal_state = TRUE
)

ps2 <- modify_transition(
  ps1,
  list(km_limit = c(20, 20))
  )
}
