% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{norm}
\alias{multinom}
\alias{norm}
\title{Convenience Probability Density Functions for 
Probabilistic Analysis}
\usage{
norm(mean, sd)

multinom(...)
}
\arguments{
\item{mean}{Distribution mean.}

\item{sd}{Distribution standard deviation.}

\item{...}{Dirichlet distribution parameters.}
}
\description{
Convenience Probability Density Functions for 
Probabilistic Analysis
}

