% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{normal}
\alias{lognormal}
\alias{make_gamma}
\alias{multinom}
\alias{normal}
\alias{prop}
\title{Convenience Probability Density Functions for 
Probabilistic Analysis}
\usage{
normal(mean, sd)

lognormal(mean, sd)

make_gamma(mean, sd)

prop(prob, size)

multinom(...)
}
\arguments{
\item{mean}{Distribution mean.}

\item{sd}{Distribution standard deviation.}

\item{prob}{Proportion.}

\item{size}{Size of sample used to estimate proportion.}

\item{...}{Dirichlet distribution parameters.}
}
\description{
Convenience Probability Density Functions for 
Probabilistic Analysis
}

