% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{normalize_ce}
\alias{normalize_ce}
\title{Normalize Cost and Effect}
\usage{
normalize_ce(x)
}
\arguments{
\item{x}{Result of \code{run_model}.}
}
\value{
Input with normalized \code{.cost} and
  \code{.effect}, ordered by \code{.effect}.
}
\description{
Normalize cost and effect values taking base model as a
reference.
}

