% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p}
\alias{p}
\title{Ensure Value is a Correct Probability}
\usage{
p(x, warn = FALSE)
}
\arguments{
\item{x}{A probability vector that may take values
outside the 0-1 range}

\item{warn}{Throw warnings when values are corrected?}
}
\value{
A correct probability vector.
}
\description{
Ensure the value returned is a correct probability
between 0 and 1.
}
\details{
Values less than 0 are set to 0, values more than 1 are
set to 1.
}
\examples{

p(c(-.1, 0, .5, 1, 1.1))
}
