% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{parse_state_file_info}
\alias{parse_state_file_info}
\title{Parse state file info files}
\usage{
parse_state_file_info(state_file_info)
}
\arguments{
\item{state_file_info}{information read from a state file}
}
\value{
a list with three components:  parsed state information,
discount rates, and the count type to be used in computing
each value
}
\description{
Parse state file info files
}
