% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{part_survs_from_surv_inputs}
\alias{part_survs_from_surv_inputs}
\title{Convert saved fits to partitioned survival objects}
\usage{
part_survs_from_surv_inputs(surv_inputs, state_names)
}
\arguments{
\item{surv_inputs}{a list of matrices of \code{flexsurvreg} objects,
for example the first element of the output of \code{survival_from_data}.}

\item{state_names}{names of states of the model}
}
\value{
a matrix of partitioned survival objects.    The rows
correspond to different distributions used in survival fitting,
the columns to different subsets of the data.
}
\description{
Convert saved fits to partitioned survival objects
}
\details{
\code{surv_inputs} must have one element with "OS" or
"os" in the name, containing the matrix of fits for overall survival,
and another element with "PFS" or "pfs" in the name, containing
the matrix of fits for progression-free survival.
}
