% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{partitioned_survival_from_tabular}
\alias{partitioned_survival_from_tabular}
\title{Create a partitioned survival model object from tabular input}
\usage{
partitioned_survival_from_tabular(ref, df_env, model_names)
}
\arguments{
\item{ref}{Name of the reference file.}

\item{df_env}{an environment in which survival results will be stored}

\item{model_names}{names of the models}
}
\value{
a list, with one element for each model.  Each element of
  the list is also a list, with one element for each survival
  fit (for example, one for overall survival and one for progression-free)
  survival).
}
\description{
Create a partitioned survival model object from tabular input
}

