% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot.eval_model_list}
\alias{plot.eval_model_list}
\title{Plot Results of Markov Model}
\usage{
\method{plot}{eval_model_list}(x, type = c("counts", "ce"), model = 1, cost,
  effect, ...)
}
\arguments{
\item{x}{Result from \code{\link{run_models}}.}

\item{type}{Type of plot, see details.}

\item{model}{Name or position of model of interest.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Various plots for Markov models.
}
\details{
\code{type = "counts"} represents state memberships (corrected) by cycle.
}

