% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot.eval_sensitivity}
\alias{plot.eval_sensitivity}
\title{Plot Sensitivity Analysis}
\usage{
\method{plot}{eval_sensitivity}(x, model = 1, value, xlab = "Parameter",
  ylab = value, ...)
}
\arguments{
\item{x}{A result of \code{\link{run_sensitivity}}.}

\item{model}{Name or index of model to plot.}

\item{value}{State value to plot.}

\item{xlab}{x-axis label.}

\item{ylab}{t-axis label.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the results of a sensitivity analysis as a tornado plot.
}

