% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_print.R
\name{plot.probabilistic}
\alias{plot.probabilistic}
\title{Plot Results of Probabilistic Analysis}
\usage{
\method{plot}{probabilistic}(x, type = c("ce", "ac"), values = seq(0, 1e+05,
  1000), ...)
}
\arguments{
\item{x}{Result from \code{\link{run_models}}.}

\item{type}{Type of plot, see details.}

\item{values}{Values for CEAC.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Various plots for Markov models probabilistic analysis.
}
\details{
\code{type = "ac"} plots cost-effectiveness acceptability
curves, \code{type = "ce"} plots results on the 
cost-efficiency plane.
}

