% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_print.R
\name{plot.run_models}
\alias{plot.run_models}
\title{Plot Results of a Markov Model}
\usage{
\method{plot}{run_models}(x, type = c("counts", "ce"), model = 1, ...)
}
\arguments{
\item{x}{Result from \code{\link{run_models}}.}

\item{type}{Type of plot, see details.}

\item{model}{Name or position of model of interest.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Various plots for Markov models.
}
\details{
\code{type = "counts"} represents state memberships 
(corrected) by cycle, \code{type = "ce"} plots models on
the cost-efficiency plane with the efficiency frontier.
}

