% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{plot_fit_data}
\alias{plot_fit_data}
\title{Plot fit data}
\usage{
plot_fit_data(data_to_plot, type = c("survival", "cumulative hazard"),
  logy = FALSE, scale_time = 1, time_label = "time",
  max_scaled_time = Inf, title = NULL, x_axis_gap)
}
\arguments{
\item{data_to_plot}{a data frame from
\code{\link[=prepare_plot_data_from_fit_tibble]{prepare_plot_data_from_fit_tibble()}}}

\item{type}{\code{survival} or \code{cumulative hazard}}

\item{logy}{should the \code{y} axis be on the logarithmic scale?}

\item{scale_time}{times are multiplied by this.  So if your
fit was done on a time scale of days, and you want to plot
by weeks, set this to 1/7.}

\item{time_label}{the \code{x} axis label}

\item{max_scaled_time}{maximum time, after scaling by \code{scale_time}}

\item{title}{for the plot}

\item{x_axis_gap}{distance between breaks on the x axis}
}
\value{
a \code{ggplot2} plot object
}
\description{
Plot fit data
}
