% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{plot_fit_tibble}
\alias{plot_fit_tibble}
\title{Title}
\usage{
plot_fit_tibble(fit_tibble, treatment, set_name, type, set_for_km = "all",
  B_ci = 100, ...)
}
\arguments{
\item{fit_tibble}{the output of \code{\link[=survival_fits_from_tabular]{survival_fits_from_tabular()}}.}

\item{treatment}{the treatment for which we want to plot.}

\item{set_name}{subset name for all but the Kaplan-Meier curve.}

\item{type}{PFS or OS}

\item{set_for_km}{subset name for the Kaplan-Meier curve.}

\item{B_ci}{number of simulations to use for confidence intervals
for \code{flexsurvreg} objects; becomes the \code{B} of
\code{\link[flexsurv:summary.flexsurvreg]{flexsurv::summary.flexsurvreg()}}}

\item{...}{additional parameters to pass to \code{\link[=plot_fit_data]{plot_fit_data()}}}
}
\value{
a list of two \code{ggplot} plots - one for survival and one
for cumulative hazard
}
\description{
Title
}
