% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plur}
\alias{plur}
\alias{plur_y}
\title{Returns "s" if x > 1}
\usage{
plur(x)

plur_y(x)
}
\arguments{
\item{x}{integer.}
}
\value{
\code{"s"} or \code{""}.
}
\description{
Returns "s" if x > 1
}
\keyword{internal}
