% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{prepare_for_eval}
\alias{prepare_for_eval}
\title{Create Quosures Ready to Be Evaluated
If an expression contains the symbol \code{find}, the symbol is replaced by its
value by evaluating it in the \code{top_caller_env}.}
\usage{
prepare_for_eval(
  x,
  top_eval_env = eval_env(),
  top_caller_env = caller_env(),
  replace_find = FALSE
)
}
\arguments{
\item{x}{a list of expressions or quosures}

\item{top_eval_env}{the main environment where objects are evaluated}

\item{top_caller_env}{the caller environment, usually the global environment}

\item{replace_find}{logical, should it search and replace the \code{find} symbol in
the expressions?}
}
\value{
A list of quosures with the environment \code{top_eval_environment}
}
\description{
Create Quosures Ready to Be Evaluated
If an expression contains the symbol \code{find}, the symbol is replaced by its
value by evaluating it in the \code{top_caller_env}.
}
\details{
Note that you cannot create a call different from \code{y = find(x)}.
}
\examples{
 age <- 12
 params <- define_parameters(
  age_init = find(age),
  age = age_init + model_time
 )
 prepare_for_eval(params, replace_find = TRUE)
 
\dontrun{
 ## The following code leads to an error
  params <- define_parameters(
   age = find(age) + model_time
  )
  prepare_for_eval(params, replace_find = TRUE)
 }
 
}
\seealso{
define_parameters
}
