% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{prepare_plot_data_from_fit_tibble}
\alias{prepare_plot_data_from_fit_tibble}
\title{Prepare fit data for plotting}
\usage{
prepare_plot_data_from_fit_tibble(fit_tib)
}
\arguments{
\item{fit_tib}{a tibble containing fits}
}
\value{
a tibble with the necessary data
}
\description{
Prepare fit data for plotting
}
