% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{projectFn}
\alias{projectFn}
\title{Project Beyond a Survival Distribution with Another (pairwise)}
\usage{
projectFn(dist1, dist2_list)
}
\arguments{
\item{dist1}{Survival distribution to project from.}

\item{dist2_list}{A list containing distribution to project with and the
time at which projection begins.}
}
\value{
A \code{surv_projection} object.
}
\description{
Project survival from a survival distribution using another survival
distribution at the specified cutpoint.  Used by project to reduce
the list of distributions.
}
\keyword{internal}
