% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.R
\name{r_norm}
\alias{r_norm}
\title{Convenience Probability Density Functions for 
Probabilistic Analysis}
\usage{
r_norm(mean, sd)
}
\arguments{
\item{mean}{Distribution mean.}

\item{sd}{Distribution standard deviation.}
}
\value{
A function taking a probability as argument and
  returning the quantile from the specified ditribution.
}
\description{
Convenience Probability Density Functions for 
Probabilistic Analysis
}

