% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{read_file}
\alias{read_file}
\title{Read the accepted file formats for tabular input}
\usage{
read_file(file_name)
}
\arguments{
\item{file_name}{File name.}
}
\value{
A \code{data.frame}.
}
\description{
Columns starting with '.comment' are ignored.
}
\keyword{internal}

