% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_plot.R
\name{reindent_transition}
\alias{reindent_transition}
\title{Reindent Transition Matrix}
\usage{
reindent_transition(x, print = TRUE)
}
\arguments{
\item{x}{A transition matrix.}

\item{print}{Print result?}
}
\value{
The reindented matrix as a text string,
invisibly.
}
\description{
Reindent Transition Matrix
}
