% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi.R
\name{run_bcea}
\alias{run_bcea}
\title{Use the BCEA package}
\usage{
run_bcea(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=run_psa]{run_psa()}}.}

\item{...}{Additional arguemnts passed to \code{\link[BCEA:bcea]{BCEA::bcea()}}.}
}
\value{
A BCEA analysis.
}
\description{
Interfaces the output of \code{\link[=run_psa]{run_psa()}} into the BCEA package.
}
\details{
The BCEA package is needed for this function to work.
}
