% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{run_model_tabular}
\alias{run_model_tabular}
\title{Run Analyses From Files}
\usage{
run_model_tabular(
  location,
  reference = "REFERENCE.csv",
  run_dsa = TRUE,
  run_psa = TRUE,
  run_demo = TRUE,
  save = FALSE,
  overwrite = FALSE
)
}
\description{
Run Analyses From Files
}
