% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_from_files.R
\name{run_models_from_files}
\alias{run_models_from_files}
\title{Run Analyses Specified in Files}
\usage{
run_models_from_files(base_dir, ref_file, N.prob = 50, cost, effect,
  init = NULL, base_model = NULL, save_outputs = TRUE,
  overwrite = FALSE, ...)
}
\arguments{
\item{base_dir}{directory in which \code{ref-file} will 
be found.}

\item{ref_file}{name of a file specifying where model 
inputs will be found.}

\item{N.prob, }{cost, effect, init arguments to 
\code{\link{run_models}} and 
\code{\link{run_probabilistic}}}

\item{base_model}{which of the models should be the base 
model}

\item{save_outputs}{Should the outputs be saved? 
Only works if the reference file specifies an output 
directory.}

\item{overwrite}{= FALSE should the outputs be 
overwritten? Not relevant if save_outputs = FALSE.}

\item{...}{additional arguments passed to 
\code{\link{execute_models}}.}
}
\value{
an unevaluated models object

a list with outputs models and model_runs
  (always), and, if appropriate input is provided, dsa
  (deterministic sensitivity analysis), psa
  (probabilistic sensitivity analysis) and demographics 
  (results across different demographic groups).
}
\description{
This function runs a model from tabular input.
}
\details{
The reference file should have two columns, \code{data} 
and \code{file}. (Additional columns are allowed, but 
will be ignored.) "data" values must include "state", 
"tm", and "parameters", and also looks for "demographics"
and "data".  The corresponding file values give the names
of the files (still in base_dir) that contain the 
corresponding information - or, in the case of "data", 
the directory containing data frames to be loaded.

For more information see the vignette: 
\code{vignette("file-input", package = "heemod")}.
}

