% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{run_models_from_tabular}
\alias{run_models_from_tabular}
\title{Run Analyses From Files}
\usage{
run_models_from_tabular(base_dir, ref_file = "REFERENCE.csv")
}
\arguments{
\item{base_dir}{Directory where the files are located.}

\item{ref_file}{Name of the reference file.}
}
\value{
A list of evaluated models (always), and, if
  appropriate input is provided, dsa (deterministic
  sensitivity analysis), psa (probabilistic sensitivity
  analysis) and demographics (results across different
  demographic groups).
}
\description{
This function runs a model from tabular input.
}
\details{
The reference file should have two columns, \code{data} 
and \code{file}. (Additional columns are allowed, but 
will be ignored.) \code{data} values must include
\code{"state"}, \code{"tm"}, and \code{"parameters"}, and
can also include \code{"options"}, \code{"demographics"} 
and \code{"data"}.  The corresponding file values give
the names of the files (located in \code{base_dir}) that
contain the corresponding information - or, in the case
of \code{"data"}, the directory containing the tables to
be loaded.
}

