% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{run_models_tabular}
\alias{run_models_tabular}
\title{Run Analyses From Files}
\usage{
run_models_tabular(location, reference = "REFERENCE.csv",
  save_outputs = FALSE, overwrite = FALSE)
}
\arguments{
\item{location}{Directory where the files are located.}

\item{reference}{Name of the reference file.}

\item{save_outputs}{Should the outputs be saved? Only works if the reference 
file specifies an output directory.}

\item{overwrite}{Should the outputs be overwritten? Not relevant if 
\code{save_outputs} = FALSE}
}
\value{
A list of evaluated models (always), and, if 
  appropriate input is provided, dsa (deterministic 
  sensitivity analysis), psa (probabilistic sensitivity 
  analysis) and demographics (results across different 
  demographic groups).
}
\description{
This function runs a model from tabular input.
}
\details{
The reference file should have two columns, \code{data} 
and \code{file}. An optional \code{absolute_path} column 
can be added, having value \code{TRUE} where an absolute 
file path is provided. \code{data} values must include 
\code{state}, \code{tm}, and \code{parameters}, and can 
also include \code{options}, \code{demographics} and 
\code{data}.  The corresponding values in the \code{file}
column give the names of the files (located in 
\code{base_dir}) that contain the corresponding 
information - or, in the case of \code{data}, the 
directory containing the tables to be loaded.
}

