% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe-conversion}
\alias{as_integer_safe}
\alias{as_numeric_safe}
\alias{safe-conversion}
\alias{safe_convert}
\title{Safely Convert From Characters to Numbers}
\usage{
safe_convert(x, f)

as_numeric_safe(x)

as_integer_safe(x)
}
\arguments{
\item{x}{A character vector.}

\item{f}{A conversion function.}
}
\value{
A converted vector.
}
\description{
These function return an error if a conversion fails.
}
\keyword{internal}

