% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapply_matrix.R
\name{sapply_matrix}
\alias{sapply_matrix}
\title{sapply_matrix}
\usage{
sapply_matrix(this_mat, fun, ...)
}
\arguments{
\item{this_mat}{a matrix or array}

\item{fun}{function to apply to each element of \code{this_mat}}

\item{...}{additional arguments for \code{fun}}
}
\value{
a matrix or array of the same dimensions as \code{this_mat},
with each element containing the results of running \code{fun} on
the corresponding element of \code{this_mat}
}
\description{
sapply_matrix
}
