% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_outputs.R
\name{save_outputs}
\alias{save_outputs}
\title{Title Save graphs and other outputs to a designated directory}
\usage{
save_outputs(outputs, base_dir, output_dir = "output", overwrite = FALSE,
  create_dir = FALSE, use_devices = c("pdf", "png"))
}
\arguments{
\item{outputs}{from run_models_from_files}

\item{base_dir}{directory in which the analysis is being conducted}

\item{output_dir}{="output" subdirectory in which to write output}

\item{overwrite}{=FALSE if \code{output_dir} already exists, should it
be overwritten?}

\item{create_dir}{=FALSE if\code{output_dir} does not exist, should it be
created?}

\item{use_devices}{=c("pdf", "png") the set of devices to output to
(each plot will be written out to each device)}
}
\value{
NULL; written for the side effect of creating files 
  in the output directory
}
\description{
Title Save graphs and other outputs to a designated directory
}

