% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{set_covariates}
\alias{set_covariates}
\title{Set Covariates of a Survival Distribution}
\usage{
set_covariates(dist, ..., data = NULL)
}
\arguments{
\item{dist}{a survfit or flexsurvreg object}

\item{...}{Covariate values representing the group for which survival
probabilities will be generated when evaluated.}

\item{data}{A an optional data frame repersenting multiple sets of covariate
values for which survival probabilities will be generated.  Can be used to
generate aggregate survival for a heterogenous set of subjects.}
}
\value{
A \code{surv_model} object.
}
\description{
Set the covariate levels of a survival model to be represented in survival
projections.
}
\examples{

fs1 = flexsurvreg(Surv(rectime, censrec)~group, data=bc)
good_model = set_covariates(group="good")
cohort = data.frame(group=c("Good","Good","Medium", "Poor"))
mixed_model = set_covariates(data=cohort)
}
