% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim.R
\name{single_strategy_cost}
\alias{single_strategy_cost}
\title{Calculate the total cost for a strategy}
\usage{
single_strategy_cost(trt_weights, trt_costs, cost_included)
}
\arguments{
\item{trt_weights}{vector, length of treatments}

\item{trt_costs}{matrix of costs - one row per treatment,
one column per cost}

\item{cost_included}{vector, length of costs - should each
cost be included}
}
\value{
total cost for a strategy
}
\description{
Calculate the total cost for a strategy
}
\details{
This function calculates the total cost for a given strategy.
  A strategy includes the proportion of different treatments, along
  with their costs.  This function also includes the ability to turn
  on or off different cost components.
}

