% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim.R
\name{single_strategy_cost_over_time}
\alias{single_strategy_cost_over_time}
\title{cost of strategy over time}
\usage{
single_strategy_cost_over_time(trt_weights_over_time, trt_costs_over_time,
  cost_included, num_patients)
}
\arguments{
\item{trt_weights_over_time}{a matrix, with one row for each strategy
and one column for each time}

\item{trt_costs_over_time}{an array; first dimension corresponds to strategies,
second dimension to costs, third to time points}

\item{cost_included}{vector, length of costs}

\item{num_patients}{number of patients in overall population}
}
\value{
vector of costs over time for the strategy
}
\description{
cost of strategy over time
}

