% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_summary.R
\name{summary.run_model}
\alias{summary.run_model}
\title{Summarise Markov Model Results}
\usage{
\method{summary}{run_model}(object, threshold = NULL, ...)
}
\arguments{
\item{object}{Output from \code{\link{run_model}}.}

\item{threshold}{ICER threshold (possibly several) for
net monetary benefit computation.}

\item{...}{additional arguments affecting the summary 
produced.}
}
\value{
A \code{summary_run_model} object.
}
\description{
Summarise Markov Model Results
}

