% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_summary.R
\name{summary.run_models}
\alias{summary.run_models}
\title{Summarise Markov Model Results}
\usage{
\method{summary}{run_models}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{run_models}}.}

\item{...}{additional arguments affecting the summary 
produced.}
}
\value{
A \code{summary_eval_model_list} object.
}
\description{
Summarise Markov Model Results
}

