% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{summary.surv_shift}
\alias{summary.surv_shift}
\title{Summarize surv_shift objects}
\usage{
\method{summary}{surv_shift}(object, summary_type = c("plot", "standard"), ...)
}
\arguments{
\item{object}{a \code{surv_shift} object}

\item{summary_type}{"standard" or "plot" - "standard"
for the usual summary of a \code{survfit} object,
"plot" for a fuller version}

\item{...}{other arguments}
}
\value{
A summary.
}
\description{
Summarize surv_shift objects
}
