% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{survival_from_data}
\alias{survival_from_data}
\title{Title Get survival analysis curves from data}
\usage{
survival_from_data(location, survival_specs, use_envir = df_env,
  dists = dists, save_fits = TRUE, just_load = FALSE,
  set_definitions = NULL)
}
\arguments{
\item{location}{base directory for the analysis.}

\item{survival_specs}{a data.frame containing information
on how to perform each fit - see @details.}

\item{use_envir}{An environment for the results to be saved in.}

\item{dists}{the distributions to use to fit the survival function}

\item{save_fits}{should fits be saved to disk?  Can be useful for testing.}

\item{just_load}{If TRUE, data files are ignored in favor of loading fits
from \code{fit_files}}

\item{set_definitions}{definitions of different subsets to fit}
}
\value{
A list with two elements:  \itemize{
\item{\code{best_models},
a list with the fits for each data file passed in; and}
\item{\code{envir},
an environment containing the models so they can be referenced to
get probabilities.}
}
}
\description{
Title Get survival analysis curves from data
}
\details{
By default, the function fits with seven
different distribution fucntions:
exponential,  Weibull,  lognormal, log-logistic,
Gompertz, gamma, and generalized gamma.

survival_specs contains information about how to create the fits:
the directory (or directories) in which
files are located, the file names, the names of the columns
that should be used for time, events, and censorship status,
and where and under what names to save the fits.
If data_files is NULL and fit files exists, then fit_files
should have the names of files in which survival models are kept.
If data_files is not NULL, then survival models will be fit from
the data in data_files (using the \code{flexsurvreg} package), and if fit_files
is also not NULL, the survival models will be saved in fit_files.

If \code{data_directory} begins with "\", with the platform file separator
(\code{.Platform$file.sep}) or with a letter and a colon (to accomodate Windows),
it's considered absolute and not appended to \code{location}.
}
