% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{test_model_name_match}
\alias{test_model_name_match}
\title{Throw an error if state names and model names don't match}
\usage{
test_model_name_match(state_model_names, transition_model_names)
}
\arguments{
\item{state_model_names}{names of models defined in state specification}

\item{transition_model_names}{names of models defined in transition specification}
}
\value{
an error with an informative message if the two inputs are
different; otherwise nothing
}
\description{
Throw an error if state names and model names don't match
}
