% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{trans_probs_from_surv}
\alias{compute_counts_part_surv}
\alias{trans_probs_from_surv}
\alias{trans_probs_from_surv.flexsurvreg}
\alias{trans_probs_from_surv.list}
\title{Transition probabilities from a survival model.}
\usage{
trans_probs_from_surv(x, ...)

\method{trans_probs_from_surv}{flexsurvreg}(surv_fit, use_km_until = 0,
  markov_cycle, markov_cycle_length = 1, pred_type = c("prob", "surv"))

\method{trans_probs_from_surv}{list}(dist_list, use_km_until, markov_cycle,
  markov_cycle_length = 1, pred_type = c("prob", "surv"))

compute_counts_part_surv(part_surv_obj, use_km_until, num_patients,
  markov_cycle, markov_cycle_length, state_names)
}
\arguments{
\item{x}{Either an object of \code{\link[flexsurv]{flexsurvreg}} 
or a list specifying a distribution and its arguments.}

\item{...}{arguments to be passed through by UseMethod}

\item{surv_fit}{An object of class \code{\link[flexsurv]{flexsurvreg}} 
that gives    the parametric survival fit object.}

\item{use_km_until}{Up to what time should Kaplan-Meier estimates be used?
Model predictions will be used thereafter.
See \code{Details}.}

\item{markov_cycle}{The Markov cycles for which to predict.}

\item{markov_cycle_length}{The value of a Markov cycle in absolute time units.}

\item{pred_type}{either \code{prob}, for transition probabilities, or 
\code{surv}, for survival}

\item{dist_list}{A list object with \code{dist_name} and other elements
giving appropriate arguments to \code{dist_name}}

\item{part_surv_obj}{A \code{flexsurvreg} object, or a list
specifying a distribution.}

\item{num_patients}{The number of patients being modeled.  Survival
probabilities are between 0 and 1, so we multiply up by num_patients.}

\item{state_names}{Name of the states to be assigned to the counts.}

\item{use_km_until}{to what time should Kaplan-Meier estimates be used?
Model predictions will be used thereafter.}
}
\value{
A \code{cycle_counts} object.
}
\description{
Transition probabilities from a survival model.

Given a partitioned survival model, returns the number of individuals 
per state per cycle.
}
\details{
The transition probabilities are estimated from a parametric fit to survival data. 
	This function is to be used when specifying transition probabilities based on 
	a survival model.

	Kaplan-Meier probabilities will be used until time \code{use_km_until}.	
	For example: if \code{use_km_until = 10},
	then KM estimates will be used for Markov cycles up to number 9,
	and model predictions	would be used thereafter.
	
	If \code{use_km_until = 0}, then only model probabilities will be used.
	\code{use_km_until < 0} will produce an error.
	
	\code{use_km_until} is used only when an object of class \code{flexsurvreg}
	  is passed to the function, because when only a distribution with arguments
	  is passed, there is no Kaplan-Meier curve.
  @return
  Returns the Markov transition probability for the current Markov cycle,
		as estimated by the Kaplan-Meier curve or from a parametric fit to 
		survival data, as specified by use_km_until.
		@examples 
		trans_probs_from_surv(list(dist_name = "exp", rate = 0.01))

See \code{trans_probs_from_surv} for further information on
  \code{part_surv_obj}, \code{use_km_until}, and the \code{markov_cycle}
  arguments.
}
\keyword{internal}

