% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ae_costs.R
\name{weighted_ae_costs}
\alias{weighted_ae_costs}
\alias{weighted_ae_costs_all}
\title{Calculate weighted costs associated with adverse events}
\usage{
weighted_ae_costs(ae_df, treatment = NULL, treatment_col = "treatment",
  prob_col = "prob", cost_col = "cost")

weighted_ae_costs_all(ae_df, treatment_col = "treatment", ...)
}
\arguments{
\item{ae_df}{\code{data.frame} with information about adverse events}

\item{treatment}{the treatment for which the adverse event costs 
should be calculated.}

\item{treatment_col}{column of \code{ae_df} holding treatment name}

\item{prob_col}{column of \code{ae_df} holding probability of an adverse
event}

\item{cost_col}{column of \code{ae_df} holding costs}

\item{...}{arguments to be passed through from \code{weighted_ae_costs_all}
to \code{weighted_ae_costs}}
}
\value{
the weighted cost of adverse events for the given treatment
}
\description{
Calculate weighted costs associated with adverse events
}
\details{
\code{weighted_ae_costs_all} calculates the costs for each treatment
    in \code{ae_df}.
}
\examples{
ae_df <- data.frame(treatment = rep(c("T1", "T2", "T3"), each = 3),
ae = rep(c("AE1", "AE2", "AE3"), 3),
prob = c(0.25, 0, 0.5, 0.2, 0.2, 0.2, 0, 0, 0),
cost = rep(c(1000, 500, 200), 3))
weighted_ae_costs(ae_df, treatment = "T2")
ae_df <- data.frame(treatment = rep(c("T1", "T2", "T3"), each = 3),
                    ae = rep(c("AE1", "AE2", "AE3"), 3),
                    prob = c(0.25, 0, 0.5, 0.2, 0.2, 0.2, 0, 0, 0),
                    cost = rep(c(1000, 500, 200), 3))
weighted_ae_costs_all(ae_df)
}

