% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gho_mortality.R
\name{who-mortality}
\alias{get_who_mr}
\alias{get_who_mr_}
\alias{who-mortality}
\title{Use WHO Mortality Rate}
\usage{
get_who_mr_(age, sex, country, year = "latest", pool = FALSE)

get_who_mr(age, sex, country, year = "latest", pool = FALSE)
}
\arguments{
\item{age}{age as a continuous variable.}

\item{sex}{sex as \code{"FMLE"}-\code{"MLE"},
\code{0}-\code{1} (male = 0, female = 1) or
\code{1}-\code{2} (male = 1, female = 2).}

\item{country}{Country code (see details).}

\item{year}{Use data from that year. Defaults to 
\code{"latest"}.}

\item{pool}{Pool female and male mortality rates?}
}
\value{
This function should be used within 
  \code{\link{define_matrix}} or 
  \code{\link{define_parameters}}.
}
\description{
Returns age and sex-specific mortality probabilities for 
a given country.
}
\details{
The results of \code{get_who_mr} are memoised for 1 hour 
to increase resampling performance. \code{get_who_mr_} is
not memoised.
}
\examples{

define_matrix(
  C, get_who_mr(age = 50 + markov_cycle, sex = "FMLE", country = "FRA"),
  0, 1
)

}

