% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit_output.R
\name{write_fits_to_excel_from_tibble}
\alias{write_fits_to_excel_from_tibble}
\title{Write fit information to an Excel workbook.}
\usage{
write_fits_to_excel_from_tibble(fit_tibble, wb, skip_at_start = 3,
  skip_between = 1, alignment = c("horizontal", "vertical"))
}
\arguments{
\item{fit_tibble}{a tibble of fits}

\item{wb}{the workbook object, or a character string with its file path}

\item{skip_at_start}{how many lines to skip at the top of each page}

\item{skip_between}{how many lines to skip between output sections}

\item{alignment}{either "horizontal" or "vertical"}
}
\value{
data suitable for plotting, invisibly.
}
\description{
Write fit information to an Excel workbook.
}
\details{
Called for the side effect of creating Excel output.

\code{alignment = 'vertical'} means sections are written one after
the other down the page.   If \code{alignment = 'horizontal'}, then
some parts will be written next to each other.
}
