% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wtd_summary}
\alias{wtd_summary}
\title{Weighted Summary}
\usage{
wtd_summary(x, weights = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A vector of weights, same length as 
\code{x}.}
}
\value{
A vector with values \code{Min., 1st Qu., Median,
  Mean, 3rd Qu., Max.}.
}
\description{
Compute a weighted summary of a numeric vector.
}
\details{
If \code{weights} is \code{NULL} an unweighted summar is
returned.
}
\keyword{internal}

