% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{calc_prob_from_surv}
\alias{calc_prob_from_surv}
\title{Calculate Probability of Event}
\usage{
calc_prob_from_surv(x)
}
\arguments{
\item{x}{A vector of conditional event probabilities.}
}
\value{
The per-cycle event probabilities.
}
\description{
Calculates the per-cycle event probabilities from a vector of survival
probabilities.
}
\keyword{internal}
