% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_eval.R
\name{check_cycle_inputs}
\alias{check_cycle_inputs}
\title{Check Cycle and Time Inputs}
\usage{
check_cycle_inputs(cycle, cycle_length)
}
\arguments{
\item{cycle}{The \code{markov_cycle} or
\code{state_time} for which to predict.}

\item{cycle_length}{The value of a Markov cycle in
absolute time units.}
}
\description{
Performs checks on the cycle and time inputs to \code{eval_surv}
}
\keyword{internal}
