% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_eval.R
\name{check_matrix}
\alias{check_matrix}
\title{Check Markov Model Transition Matrix}
\usage{
check_matrix(x)
}
\arguments{
\item{x}{a matrix.}
}
\value{
\code{NULL}
}
\description{
Check whether a matrix fullfills the conditions to be a 
transition matrix.
}
\details{
This function is called by \code{\link{eval_matrix}} and 
should not be used directly.

Checks whether all rows sum to 1 and all probabilities 
are between 0 and 1.
}
\keyword{internal}

