% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine_models}
\alias{combine_models}
\title{Combine Multiple Models}
\usage{
combine_models(newmodels, weights, oldmodel)
}
\arguments{
\item{newmodels}{A list of models run over a set of 
multiple parameters.}

\item{weights}{A vector of weigths, same length as the
number of parameter sets.}

\item{oldmodel}{The original model.}
}
\value{
A \code{combined_models} object, mostly similar
  to a result from \code{\link{run_model}}. \code{plot}
  and \code{summary} methods are available.
}
\description{
Given a set of models run with different parameters, 
return aggregated results to estimate population-level
values.
}
\keyword{internal}

