% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{compute_counts_part_surv}
\alias{compute_counts_part_surv}
\title{Compute Counts for Partitioned Survival Models}
\usage{
compute_counts_part_surv(part_surv_obj, use_km_until, num_patients,
  markov_cycle, markov_cycle_length, state_names)
}
\arguments{
\item{num_patients}{The number of patients being modeled.}

\item{state_names}{Name of the states to be assigned to 
the counts.}

\item{x}{A list of \code{flexsurvreg} objects, or a list
of lists specifying distributions.}

\item{km_limit}{To what time should Kaplan-Meier 
estimates be used? Model predictions will be used 
thereafter.}
}
\value{
A \code{cycle_counts} object.
}
\description{
Compute count of individual in each state per cycle based
on a partitioned survival model.
}
\details{
The elements of \code{x} should be called \code{pfs} and 
\code{os} for progression-free survival and overall
survival respectively.

See \code{\link{get_probabilities}} for further 
information on \code{km_limit}, and the
\code{markov_cycle} arguments.
}
\keyword{internal}

