% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{compute_icer}
\alias{compute_icer}
\title{Compute ICER}
\usage{
compute_icer(x, cost, effect)
}
\arguments{
\item{x}{Result of \code{\link{run_models}}}

\item{cost}{character. Variable name corresponding to 
cost.}

\item{effect}{character. Variable name corresponding to
efficicacy (or utility).}
}
\value{
An object of class \code{mat_icer}.
}
\description{
Compute ICER for Markov models.
}

