% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_summary.R
\name{compute_icer}
\alias{compute_icer}
\title{Compute ICER}
\usage{
compute_icer(x, model_order = order(x$.effect))
}
\arguments{
\item{x}{Result of \code{\link{run_models}}.}

\item{model_order}{Order in which the models should be
sorted. Default: by increasing effect.}
}
\value{
A \code{data.frame} with computed ICER.
}
\description{
Compute ICER for Markov models.
}
\details{
Models are ordered by effectiveness and ICER are computed
sequencially.
}
\keyword{internal}

