% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_eval.R
\name{compute_values}
\alias{compute_values}
\title{Compute State Values per Cycle}
\usage{
compute_values(states, counts)
}
\arguments{
\item{states}{An object of class \code{eval_state_list}.}

\item{counts}{An object of class \code{cycle_counts}.}
}
\value{
A data.frame of state values, one column per 
  state value and one row per cycle.
}
\description{
Given states and counts, computes the total state values 
per cycle.
}
\keyword{internal}

