% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_inflow}
\alias{define_inflow}
\alias{define_inflow_}
\title{Define Inflow for a BIA}
\usage{
define_inflow(...)

define_inflow_(.dots)
}
\arguments{
\item{...}{Name-value pairs of expressions defining
inflow counts.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object similar to the return value of
\code{\link[=define_parameters]{define_parameters()}}.
}
\description{
Define Inflow for a BIA
}
