% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_define.R
\name{define_multinom}
\alias{define_multinom}
\title{Define That Parameters Belong to the Same Multinomial 
Distribution}
\usage{
define_multinom(x)
}
\arguments{
\item{x}{A vector of parameter names.}
}
\value{
An object of class \code{multinomial}.
}
\description{
Define That Parameters Belong to the Same Multinomial 
Distribution
}

