% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{define_part_surv}
\alias{define_part_surv}
\alias{define_part_surv_}
\title{Define Partitioned Survival}
\usage{
define_part_surv(..., state_names, terminal_state = FALSE)

define_part_surv_(.dots, state_names)
}
\arguments{
\item{state_names}{named character vector, length 3 or 4.
State names for progression-free state, progression, 
(optionally terminal) and death respectively. Elements 
should be named \code{"progression_free"}, 
\code{"progression"}, (optionally \code{"terminal"}), 
and \code{"death"}. See examples.}

\item{terminal_state}{Should a terminal state be 
included? Only used when state names are not provided.}

\item{.dots}{Used for NSE, lazy dots with names
\code{pfs} and \code{os}.}

\item{pfs}{Result of \code{\link{get_prob_from_surv}} for
progression-free survival.}

\item{os}{Idem for overall survival.}
}
\value{
A \code{part_surv} object.
}
\description{
Define a partitioned survival model with progression-free
survival and overall survival.
}
\examples{
probs_pfs <- define_survival("exp", rate = 1)
probs_os <- define_survival("exp", rate = .5)

define_part_surv(
  pfs = get_probs_from_surv(probs_pfs, cycle = markov_cycle),
  os = get_probs_from_surv(probs_os, cycle = markov_cycle),
  state_names = c(
    progression_free = "A",
    progression = "B",
    terminal = "C",
    death = "D"
  )
)
#' identical to:
define_part_surv(
  pfs = get_probs_from_surv(probs_pfs, cycle = markov_cycle),
  os = get_probs_from_surv(probs_os, cycle = markov_cycle),
  terminal_state = TRUE
)

}

