% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{define_starting_values}
\alias{define_starting_values}
\alias{define_starting_values_}
\title{Define Starting State Values}
\usage{
define_starting_values(...)

define_starting_values_(.dots)
}
\arguments{
\item{...}{Name-value pairs of expressions defining
starting values. The names must correspond to an existing state value.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object similar to the return value of
\code{\link[=define_parameters]{define_parameters()}}.
}
\description{
This function is meant to be used inside \code{\link[=define_strategy]{define_strategy()}} and
\code{\link[=define_state]{define_state()}}.
}
\details{
The behaviour is different following the function using \code{\link[=define_starting_values]{define_starting_values()}}
as an argument.
\itemize{
\item When used inside \code{\link[=define_strategy]{define_strategy()}}, the state values are modified for the
first cycle in each state
\item When used inside \code{\link[=define_state]{define_state()}}, the state values are modified for counts
entering the state
}
}
