% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{define_survival}
\alias{define_survival}
\title{Define a Survival Distribution}
\usage{
define_survival(distribution = c("exp", "weibull", "lnorm", "gamma",
  "gompertz", "gengamma"), ...)
}
\arguments{
\item{distribution}{A parametric survival distribution.}

\item{...}{Additional distribution parameters (see xxx).}
}
\value{
A \code{surv_dist} object.
}
\description{
Define a parametric survival distribution, to be used
by \code{\link{get_probs_from_surv}}.
}
\examples{

define_survival(distribution = "exp", rate = .5)
define_survival(distribution = "gompertz", rate = .5, shape = 1)

}

