% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discount}
\alias{discount}
\title{Discount a Quantity Over Time}
\usage{
discount(x, r, first = FALSE)
}
\arguments{
\item{x}{numeric. A quantity to discount.}

\item{r}{discount rate.}

\item{first}{logical. Should the discouting start at the
first value ?}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
This function should only take as an \code{x} argument 
the names of variables already defined in 
\code{\link{define_parameters}} or 
\code{\link{define_state}}, and not numeric constants.
}
\examples{

discount(rep(10, 5), .02)
discount(rep(10, 5), .02, first = FALSE)

}
\keyword{internal}

