% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_eval.R
\name{eval_resample}
\alias{eval_resample}
\title{Evaluate Resampling Definition}
\usage{
eval_resample(resample, N)
}
\arguments{
\item{resample}{A \code{\link{define_psa}} object.}

\item{N}{> 0. Number of simulation to run.}
}
\value{
A \code{data.frame} of resampled values with on 
  column per parameter and \code{N} rows.
}
\description{
Evaluate Resampling Definition
}
\keyword{internal}

