% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{f_add_surv_fit_metrics}
\alias{f_add_surv_fit_metrics}
\title{Calculate additional metrics to evaluate fit of survival model.}
\usage{
f_add_surv_fit_metrics(surv_fits, metrics = c("BIC", "m2LL"))
}
\arguments{
\item{surv_fits}{A list object from \code{\link[=f_fit_survival_models]{f_fit_survival_models()}} that gives
a collection (list) of \code{flexsurvreg} parametric survival fit object.}

\item{metrics}{Metrics to calculate.}
}
\value{
A list object of parametric survival fits, containing additional fields for
the calculated fit metrics.
}
\description{
Calculate additional metrics to evaluate fit of survival model.
}
\details{
Currently calculates only:
\itemize{\item Bayesian information criterion (BIC)
\item -2*log likelihood (-2LL)}  (Objects come with AIC already calculated.)
}
