% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{find_scaled_doses}
\alias{find_scaled_doses}
\title{Find the total dose, given a scaling parameter
(e.g. find total dose for a 70kg patient for a 2mg/kg drug dosing)}
\usage{
find_scaled_doses(doses, dosing_units, scaling, scaling_units)
}
\arguments{
\item{doses}{a vector of doses (e.g. 2, 5, 10)}

\item{dosing_units}{a vector of units corresponding to the doses
(e.g. "mg/kg", "mg/m2", "mg")}

\item{scaling}{a vector of scaling numbers (e.g. 70, 85, 1)}

\item{scaling_units}{a vector of units for the scaling numbers
(e.g. "kg", "m2", "mg")}
}
\value{
Returns a vector of scale-adjusted doses (e.g. 140mg, 425mg, 10mg)
}
\description{
Find the total dose, given a scaling parameter
(e.g. find total dose for a 70kg patient for a 2mg/kg drug dosing)
}
