% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{gather_model_info}
\alias{gather_model_info}
\title{Gather Information for Running a Model From Tabular Data}
\usage{
gather_model_info(base_dir, ref_file)
}
\arguments{
\item{base_dir}{Directory where the files are located.}

\item{ref_file}{Name of the reference file.}
}
\value{
A list with elements: \itemize{ \item models (of 
  type \code{uneval_model}, created by 
  \code{\link{create_model_list_from_tabular}}) \item 
  param_info  \item output_dir where to store output 
  files, if specified \item demographic_file a table for 
  demographic analysis \item model_options a list of 
  model options.}
}
\description{
Gather Information for Running a Model From Tabular Data
}
\keyword{internal}

