% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R, R/run_model_define.R
\name{get_counts.updated_model}
\alias{get_counts}
\alias{get_counts.combined_model}
\alias{get_counts.eval_strategy}
\alias{get_counts.list}
\alias{get_counts.run_model}
\alias{get_counts.updated_model}
\title{Get State Membership Counts}
\usage{
\method{get_counts}{updated_model}(x, ...)

\method{get_counts}{combined_model}(x, ...)

get_counts(x, ...)

\method{get_counts}{run_model}(x, ...)

\method{get_counts}{eval_strategy}(x, ...)

\method{get_counts}{list}(x, ...)
}
\arguments{
\item{x}{Result from \code{\link{run_model}}.}

\item{...}{further arguments passed to or from other
methods.}
}
\value{
A data frame of counts per state.
}
\description{
Given a result from \code{\link{run_model}}, return 
state membership counts for a specific strategy.
}

