% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_frontier.R
\name{get_frontier}
\alias{get_frontier}
\title{Return Efficiency Frontier}
\usage{
get_frontier(x)
}
\arguments{
\item{x}{An \code{eval_model_list} object.}
}
\value{
A vector of model names on the efficiency
  frontier.
}
\description{
Return Efficiency Frontier
}
\keyword{internal}

