% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_define.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Return parameters names}
\usage{
get_parameter_names(x)
}
\arguments{
\item{x}{\code{uneval_parameters} or 
\code{eval_parameters} object.}
}
\value{
A character vector of parameter names.
}
\description{
Extract parameters names from an \code{uneval_parameters}
or \code{eval_parameters} object.
}
\keyword{internal}

