% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_probs_from_surv}
\alias{get_probs_from_surv}
\alias{get_probs_from_surv.flexsurvreg}
\alias{get_probs_from_surv.list}
\title{Extract Transition Probabilities from a Survival Model}
\usage{
get_probs_from_surv(x, ...)

\method{get_probs_from_surv}{flexsurvreg}(x, cycle, km_limit = 0,
  cycle_length = 1, type = c("prob", "surv"))

\method{get_probs_from_surv}{list}(x, cycle, km_limit = 0, cycle_length = 1,
  type = c("prob", "surv"))
}
\arguments{
\item{x}{Either a result from 
\code{\link[flexsurv]{flexsurvreg}} or a list 
specifying a distribution and its arguments.}

\item{...}{arguments to be passed to methods.}

\item{cycle}{The \code{markov_cycle} or 
\code{state_cycle} for which to predict.}

\item{km_limit}{Up to what time should Kaplan-Meier 
estimates be used? Model predictions will be used 
thereafter. See \code{Details}.}

\item{cycle_length}{The value of a Markov cycle in 
absolute time units.}

\item{type}{either \code{prob}, for transition 
probabilities, or \code{surv}, for survival}
}
\value{
Returns the Markov transition probability for the
  cycles.
}
\description{
Get probabilities from fitted survival models, to be used
by a transition matrix or a partitioned survival model.
}
\details{
If \code{use_km_until = 0}, then only model probabilities
will be used.
}
\examples{
		
trans_probs_from_surv(list(dist_name = "exp", rate = 0.01))

}

