% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R, R/run_model_define.R
\name{get_values.updated_model}
\alias{get_values}
\alias{get_values.combined_model}
\alias{get_values.eval_strategy}
\alias{get_values.list}
\alias{get_values.run_model}
\alias{get_values.updated_model}
\title{Get Strategy Values}
\usage{
\method{get_values}{updated_model}(x, ...)

\method{get_values}{combined_model}(x, ...)

get_values(x, ...)

\method{get_values}{run_model}(x, ...)

\method{get_values}{eval_strategy}(x, ...)

\method{get_values}{list}(x, ...)
}
\arguments{
\item{x}{Result from \code{\link{run_model}}.}

\item{...}{further arguments passed to or from other
methods.}
}
\value{
A data frame of values per state.
}
\description{
Given a result from \code{\link{run_model}}, return 
cost and effect values for a specific strategy.
}

