% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heemod.R
\docType{package}
\name{heemod}
\alias{heemod}
\alias{heemod-package}
\title{heemod: Health Economic Evaluation MODelling}
\description{
\code{heemod} is an R toolset for health economic 
evaluation modelling. It aims to provide a simple and 
consistent interface for Markov models specification and
comparison. Non-homogeneous Markov models (with time
varying properties) are supported.
}

