% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_from_files.R
\name{import_models_from_files}
\alias{import_models_from_files}
\title{Read Models Specified by Files}
\usage{
import_models_from_files(ref, base_model)
}
\arguments{
\item{ref}{Imported reference file.}

\item{base_model}{Which of the models should be the base 
model.}
}
\value{
A list of unevaluated models.
}
\description{
Read Models Specified by Files
}

