% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_summary.R
\name{normalize_ce}
\alias{normalize_ce}
\title{Normalize Cost and Effect}
\usage{
normalize_ce(x)
}
\arguments{
\item{x}{Result of \code{\link{run_models}} or 
\code{\link{run_probabilistic}}.}
}
\value{
Input with normalized \code{.cost} and 
  \code{.effect}, ordered by \code{.effect}.
}
\description{
Normalize cost and effect values taking base model as a 
reference.
}
\keyword{internal}

