% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_surv.R
\name{part_survs_from_surv_inputs}
\alias{part_survs_from_surv_inputs}
\title{Convert saved fits to partitioned survival objects}
\usage{
part_survs_from_surv_inputs(surv_inputs, state_names)
}
\arguments{
\item{surv_inputs}{a list of matrices of \code{flexsurvreg} objects,
for example the first element of the output of \code{survival_from_data}.}

\item{state_names}{names of states of the model}
}
\value{
a tibble of partitioned survival objects, similar to the
original tibble of survival fits, with all the columns
except type and fit, and a new column part_surv.
}
\description{
Convert saved fits to partitioned survival objects
}
\details{
surv_inputs is a tibble with columns
type (PFS or OS, not case sensitive), treatment,
set_name (for data subsets),
dist (for survival distribution assumptions),
fit (for the fitted survival object) and set_def
(how the subset of data was defined, just to keep it around)
}
