% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{partitioned_survival_from_tabular}
\alias{partitioned_survival_from_tabular}
\title{Create a partitioned survival model object from tabular input}
\usage{
partitioned_survival_from_tabular(base_dir, ref_file, df_env, state_names,
  model_names)
}
\arguments{
\item{base_dir}{directory where the reference file is located}

\item{ref_file}{Name of the reference file.}

\item{df_env}{an environment in which survival results will be stored}

\item{state_names}{names of states in the models}

\item{model_names}{names of the models}
}
\value{
a matrix of partitioned survival objects.    The rows
  correspond to different distributions used in survival fitting,
  the columns to different subsets of the data.
}
\description{
Create a partitioned survival model object from tabular input
}

