% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_print.R
\name{plot.dsa}
\alias{plot.dsa}
\title{Plot Sensitivity Analysis}
\usage{
\method{plot}{dsa}(x, type = c("simple", "difference"), result = c("cost",
  "effect", "icer"), strategy = NULL, widest_on_top = TRUE,
  remove_ns = FALSE, bw = FALSE, ...)
}
\arguments{
\item{x}{A result of \code{\link{run_dsa}}.}

\item{type}{Type of plot (see details).}

\item{result}{Plot cost, effect, or ICER.}

\item{strategy}{Name or index of strategies to plot.}

\item{widest_on_top}{logical. Should bars be sorted so
widest are on top?}

\item{remove_ns}{Remove variables that are not sensitive.}

\item{bw}{Black & white plot for publications?}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the results of a sensitivity analysis as a tornado 
plot.
}
\details{
Plot type \code{simple} plots variations of single strategy 
values, while \code{difference} plots incremental values.
}

