% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot.eval_sensitivity}
\alias{plot.eval_sensitivity}
\title{Plot Sensitivity Analysis}
\usage{
\method{plot}{eval_sensitivity}(x, ...)
}
\arguments{
\item{x}{A result of \code{\link{run_sensitivity}}.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the results of a sensitivity analysis as a tornado plot.
}

