% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_print.R
\name{plot.eval_sensitivity}
\alias{plot.eval_sensitivity}
\title{Plot Sensitivity Analysis}
\usage{
\method{plot}{eval_sensitivity}(x, type = c("simple", "diff"), model = 1,
  widest_on_top = TRUE, ...)
}
\arguments{
\item{x}{A result of \code{\link{run_sensitivity}}.}

\item{type}{Type of plot (see details).}

\item{model}{Name or index of model to plot.}

\item{widest_on_top}{Logical. Should bars be sorted so widest are on top?}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the results of a sensitivity analysis as a tornado 
plot.
}
\details{
Plot type \code{simple} plots variations of single model 
costs, while \code{diff} plots cost difference between 
the specified model and the reference model.
}

