% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{plot.surv_object}
\alias{plot.surv_object}
\title{Plot general survival models}
\usage{
\method{plot}{surv_object}(x, time, type = c("surv", "prob"), ...)
}
\arguments{
\item{x}{A survival object of class \code{surv_object}.}

\item{time}{Times for which to predict.}

\item{type}{Either \code{prob}, for transition probabilities,
or \code{surv}, for survival.}

\item{...}{Additional arguments to pass to \code{\link[=compute_surv]{compute_surv()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot general survival models
}
\details{
The function currently only highlights join
points that are at the top level; that is, for objects
with class \code{surv_projection}.

To avoid plotting the join points, set join_size to a
negative number.
}
