% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{r_norm}
\alias{r_multinom}
\alias{r_norm}
\title{Convenience Probability Density Functions for 
Probabilistic Analysis}
\usage{
r_norm(mean, sd)

r_multinom(n, total)
}
\arguments{
\item{mean}{Distribution mean.}

\item{sd}{Distribution standard deviation.}

\item{n}{Number of observed case for multinomial 
distribution.}

\item{total}{Population size for multinomial
distribution.}
}
\value{
A function taking a probability as argument and 
  returning the quantile from the specified ditribution.
}
\description{
Convenience Probability Density Functions for 
Probabilistic Analysis
}

