% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{resample_surv}
\alias{resample_surv}
\alias{resample_surv_boot}
\alias{resample_surv_dist}
\title{Resample survival distribution}
\usage{
resample_surv(n)

resample_surv_boot(x)

resample_surv_dist(x, n)
}
\arguments{
\item{n}{the number of observations to generate if dist is specified or x is a \code{surv_dist}
object

The lower n is, the higher is the variability}

\item{x}{a \code{surv_object}}
}
\description{
Resample survival distribution
}
\seealso{
plot.surv_psa
}
\keyword{internal}
