% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{resample_surv}
\alias{resample_surv}
\alias{resample_surv.default}
\alias{resample_surv.surv_pooled}
\alias{resample_surv.surv_add_haz}
\alias{resample_surv.surv_dist}
\title{Resample survival distribution}
\usage{
resample_surv(x, ...)

\method{resample_surv}{default}(x, ...)

\method{resample_surv}{surv_pooled}(x, ...)

\method{resample_surv}{surv_add_haz}(x, ...)

\method{resample_surv}{surv_dist}(x, n)
}
\arguments{
\item{x}{a \code{surv_object}}

\item{n}{the number of observations to generate if dist is specified or x is a \code{surv_dist}
object

The lower n is, the higher is the variability}
}
\description{
Resample survival distribution

#' @rdname resample_surv
#' @export
resample_surv.numeric <- function(x, ...){
structure(expr(resample_surv(!!x)),
class = c("surv_psa"))

}
}
\examples{
surv <- define_surv_dist("exp", rate = 0.5)

## Should return a similar distribution
resample_surv(surv, n = 100)

library(survival)
sf <- survfit(Surv(time, status) ~ 1, data = colon)
resample_surv(sf)
}
\seealso{
plot.surv_psa
}
