% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_input.R
\name{save_outputs}
\alias{save_outputs}
\title{Save Model Outputs}
\usage{
save_outputs(outputs, output_dir, overwrite)
}
\arguments{
\item{outputs}{Result from
\code{\link{run_model_tabular}}.}

\item{output_dir}{Subdirectory in which to write output.}

\item{overwrite}{Should the outputs be overwritten?}
}
\value{
\code{NULL}. Used for its side effect of creating
  files in the output directory.
}
\description{
Save Model Outputs
}
\keyword{internal}

