% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{summary.eval_model_list}
\alias{summary.eval_model_list}
\title{Summarise Markov Model Results}
\usage{
\method{summary}{eval_model_list}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{run_models}}.}

\item{...}{additional arguments affecting the summary
produced.}
}
\value{
A \code{summary_eval_model_list} object.
}
\description{
Summarise Markov Model Results
}

