% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_operations.R
\name{summary.surv_shift}
\alias{summary.surv_shift}
\title{Summarize surv_shift objects}
\usage{
\method{summary}{surv_shift}(object, ...)
}
\arguments{
\item{object}{a \code{surv_shift} object}

\item{...}{other arguments}
}
\description{
Summarize surv_shift objects
}
