% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{survival_fits_from_tabular}
\alias{survival_fits_from_tabular}
\alias{survival_fits_from_ref_struc}
\alias{partitioned_survival_from_tabular}
\alias{partitioned_survival_from_ref_struc}
\title{Create a partitioned survival model object from tabular input}
\usage{
survival_fits_from_tabular(base_dir, ref_file, ...)

survival_fits_from_ref_struc(ref, df_env = new.env(), save_fits = FALSE,
  just_load = FALSE)

partitioned_survival_from_tabular(base_dir, ref_file, df_env, state_names,
  save_fits = FALSE, just_load = FALSE)

partitioned_survival_from_ref_struc(ref, df_env, state_names,
  save_fits = FALSE, just_load = FALSE)
}
\arguments{
\item{base_dir}{directory where the reference file is located}

\item{ref_file}{Name of the reference file.}

\item{...}{additional arguments to be passed to
\code{\link[=survival_fits_from_ref_struc]{survival_fits_from_ref_struc()}}}

\item{ref}{data frame read from the reference file.}

\item{df_env}{an environment in which survival results will be stored}

\item{save_fits}{should the models be saved to disk?}

\item{just_load}{should we just load the models instead of fitting?}

\item{state_names}{names of states in the strategies}
}
\value{
a tibble of partitioned survival objects
(for \code{part_surv_from_tabular})
or a list (for \code{survival_fits_from_tabular}),
with each element being a similar
tibble of flexsurvreg fits for either overall survival
or progression-free survival.
There are columns to distinguish the different distributions used
used in survival fitting, different treatments, and different subsets.
}
\description{
Create a partitioned survival model object from tabular input
}
\examples{
fit_tibble <- partitioned_survival_from_tabular(system.file("tabular/surv",
                                                            package = "heemod"), 
                                  "example_oncSpecs.csv", new.env(), 
                                  c("ProgressionFree", "Progressive", 
                                  "Terminal", "Death"), 
                                  save_fits = FALSE,
                                  just_load = FALSE)
}
