% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_fit.R
\name{survival_from_data}
\alias{survival_from_data}
\title{Make sure survival inputs have the correct formats}
\usage{
survival_from_data(base_dir, surv_dir, data_files = NULL, fit_files = NULL,
  fit_names = NULL, time_col_name = "time", censor_col_name = "status",
  treatment_col_name = "treatment", fit_metric = "AIC", best_only = TRUE,
  dists = c("exp", "weibull", "lnorm", "gamma", "gompertz", "gengamma"),
  use_envir = NULL, save_fits = TRUE, just_load = FALSE,
  set_definitions = NULL)
}
\arguments{
\item{base_dir}{base directory for the analysis.}

\item{surv_dir}{subdirectory in which survival data and fits are stored.}

\item{data_files}{files containing survival data.}

\item{fit_files}{Files containing fits.}

\item{fit_names}{name(s) to save the fits to, if they are being saved in a file}

\item{time_col_name, censor_col_name, treatment_col_name}{columns to be used in the analysis}

\item{fit_metric}{Once of AIC, BIC, or m2LL}

\item{best_only}{should only the best fit for each condition
be returned?}

\item{dists}{the distributions to use to fit the survival function}

\item{use_envir}{An environment for the results to be saved in.}

\item{save_fits}{should fits be saved to disk?  Can be useful for testing.}

\item{just_load}{If TRUE, data files are ignored in favor of loading fits
from \code{fit_files}}

\item{set_definitions}{definitions of different subsets to fit}

\item{location}{base directory for the analysis}

\item{surv_ref}{data frame with survival data information}
}
\value{
a list with elements \itemize{\item{fit_files},
\item{fit_names}, \item{surv_data_files}, \item{fit_metric}}.

Title Get survival analysis curves from data

A list with two elements:  \itemize{
\item{\code{best_models},
a list with the fits for each data file passed in; and}
\item{\code{envir},
an environment containing the models so they can be referenced to
get probabilities.}
}
}
\description{
Make sure survival inputs have the correct formats
}
\details{
For survival analysis, we need, for each condition
(frequently progression-free survival and overall survival)
at least two, and possibly three, elements represented in
the survival reference file.
\itemize{
\item{\code{surv_data_file}:  the name of the file
with the survival data}
\item{\code{fit_file}: the name of a file with fits.  If
\code{surv_data_file} is not provided, the fit will
be read in from this file; if \code{surv_data_file}
is provided, the fit will be written into this file.}
\item{\code{fit_name}:  name to give the survival fit so
that it can be referred to to extract probabilities.}
}
The names for the three fields are \code{fit_name_<label>},
\code{surv_data_file_<label>}, and \code{fit_file_<label>}, where
\code{<label>} identifies that the fields go together for a single fit.
The labels for the two conditions must be \code{pfs} and \code{os}
(or \code{PFS} and \code{OS}; they are not case sensitive).
At least one of \code{surv_data_file} and \code{fit_file}
must be provided.

By default, the function fits with six different distribution fucntions:
exponential,  Weibull,  lognormal, Gompertz, gamma, and generalized gamma.

If data_files is NULL and fit files exists, then fit_files
should have the names of files in which survival models are kept.
If data_files is not NULL, then survival models will be fit from
the data in data_files (using the \code{flexsurvreg} package), and if fit_files
is also not NULL, the survival models will be saved in fit_files.

If \code{best_only = TRUE}, then \code{best_models} will be a vector
of models, one for each group.
If \code{best_only = FALSE}, \code{best_models}
will have a matrix of models, with each row corresponding to a
distribution and each column to a group.
}
