% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_from_data}
\alias{survival_from_data}
\title{Title Get survival analysis curves from data}
\usage{
survival_from_data(base_dir, surv_dir, data_files = NULL, fit_files = NULL,
  fit_names = NULL, time_col_name = "time", censor_col_name = "status",
  treatment_col_name = "treatment", fit_metric = "AIC", best_only = TRUE,
  use_envir = NULL)
}
\arguments{
\item{base_dir}{base directory for the analysis.}

\item{surv_dir}{subdirectory in which survival data and fits are stored.}

\item{data_files}{files containing survival data.}

\item{fit_files}{Files containing fits.}

\item{fit_names}{name(s) to save the fits to, if they are being saved in a file}

\item{time_col_name, censor_col_name, treatment_col_name}{columns to be used in the analysis}

\item{fit_metric}{Once of AIC, BIC, or m2LL}

\item{best_only}{should only the best fit for each condition
be returned?}

\item{use_envir}{An environment for the results to be saved in.}
}
\value{
A list with two elements:  \itemize{
   \item{\code{best_models}, 
   a list with the fits for each data file passed in; and} 
   \item{\code{envir}, 
   an environment containing the models so they can be referenced to 
   get probabilities.}
   }
}
\description{
Title Get survival analysis curves from data
}
\details{
If data_files is NULL and fit files exists, then fit_files
  should have the names of files in which survival models are kept.
  If data_files is not NULL, then survival models will be fit from
  the data in data_files (using the \code{flexsurvreg} package), and if fit_files
  is also not NULL, the survival models will be saved in fit_files.
  
  If \code{best_only = TRUE}, then \code{best_models} will be a vector
  of models, one for each group.   
  If \code{best_only = FALSE}, \code{best_models}
  will have a matrix of models, with each row corresponding to a 
  distribution and each column to a group.
}

