% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_eval.R
\name{compute_counts_part_surv}
\alias{compute_counts_part_surv}
\title{Compute Count of Individual in Each State per Cycle based
on a partitioned survival model}
\usage{
compute_counts_part_surv(x, km_limit, num_patients, markov_cycle,
  markov_cycle_length = 1)
}
\arguments{
\item{num_patients}{The number of patients being modeled.
Survival probabilities are between 0 and 1, so we
multiply up by num_patients.}

\item{part_surv_obj}{A \code{flexsurvreg} object, or a
list specifying a distribution.}

\item{use_km_until}{to what time should Kaplan-Meier
estimates be used? Model predictions will be used
thereafter.}

\item{state_names}{Name of the states to be assigned to
the counts.}
}
\value{
A \code{cycle_counts} object.
}
\description{
Given a partitioned survival model, returns the number of
individuals per state per cycle.
}
\details{
See \code{trans_probs_from_surv} for further
  information on \code{part_surv_obj},
  \code{use_km_until}, and the \code{markov_cycle} 
  arguments.
}
\keyword{internal}

